U.Codebase.Branch.Diff

Documentation

newtype TreeDiff m

A tree of local diffs. Each node of the tree contains the definition diffs at that path.

Constructors

TreeDiff

Fields

unTreeDiff :: Cofree (Compose (Map NameSegment) m) DefinitionDiffs

hoistTreeDiff :: Functor m => (forall x. m x -> n x) -> TreeDiff m -> TreeDiff n

data NameChanges

A summary of a TreeDiff, containing all names added and removed. Note that there isn't a clear notion of a name "changing" since conflicts might muddy the notion by having multiple copies of both the from and to names, so we just talk about adds and removals instead.

Constructors

NameChanges

Fields

termNameAdds :: [(Name, Referent)]

termNameRemovals :: [(Name, Referent)]

typeNameAdds :: [(Name, Reference)]

typeNameRemovals :: [(Name, Reference)]

data DefinitionDiffs

Represents the changes to definitions at a given path, not including child paths.

Note: doesn't yet include any info on patch diffs. Feel free to add it.

Constructors

DefinitionDiffs

Fields

termDiffs :: Map NameSegment (Diff Referent)

typeDiffs :: Map NameSegment (Diff Reference)

data Diff a

Constructors

Diff

Fields

adds :: Set a

removals :: Set a

data NameBasedDiff

A name-based diff for namespaces N1 and N2 is (for both terms and types) a relation between references, where `a R b` if:

a has name n in N1, and b has the same name n in N2

a != b

Constructors

NameBasedDiff

Fields

terms :: Relation Reference Reference

types :: Relation Reference Reference

diffBranches :: Branch Transaction -> Branch Transaction -> Transaction (TreeDiff Transaction)

Diff two Branches, returning a tree containing all of the changes

allNameChanges :: Monad m => Maybe Name -> TreeDiff m -> m NameChanges

Get a summary of all of the name adds and removals from a tree diff.

The provided name will be prepended to all names in the output diff, and can be useful if diffing branches at a specific sub-tree, but you can pass Nothing if you're diffing from the root.

nameBasedDiff :: Monad m => TreeDiff m -> m NameBasedDiff

Get a NameBasedDiff from a TreeDiff.

streamNameChanges :: (Monad m, Monoid r) => Maybe Name -> TreeDiff m -> (Maybe Name -> NameChanges -> m r) -> m r

Stream a summary of all of the name adds and removals from a tree diff. Callback is passed the diff from one namespace level at a time, with the name representing that location. Accumulator is folded strictly, use () if you don't need one.