Unison.Codebase.Branch

Branch types

newtype Branch (m :: Type -> Type)

A node in the Unison namespace hierarchy along with its history.

Constructors

Branch

Fields

_history :: UnwrappedBranch m

type UnwrappedBranch (m :: Type -> Type) = Causal m (Branch0 m)

data Branch0 (m :: Type -> Type)

A node in the Unison namespace hierarchy.

_terms and _types are the declarations at this level. _children are the nodes one level below us. _edits are the Patchs stored at this node in the code.

The remaining fields are derived from the four above. None of the record fields are exported to avoid accidental tweaking without updating the associated derived fields.

Use either the lensy accessors or the field getters.

data Raw

type Star r n = Star r n

type NamespaceHash (m :: Type -> Type) = HashFor (Branch0 m)

A Hash for a namespace itself, it doesn't incorporate any history.

Branch construction

fromUnconflictedDefns :: forall (m :: Type -> Type). DefnsF (Map Name) Referent TypeReference -> Branch0 m

fromNametree :: forall (m :: Type -> Type). Nametree (DefnsF (Map NameSegment) Referent TypeReference) -> Branch0 m

branch0 :: Star Referent NameSegment -> Star TypeReference NameSegment -> Map NameSegment (Branch m) -> Map NameSegment (PatchHash, m Patch) -> Branch0 m

one :: forall (m :: Type -> Type). Branch0 m -> Branch m

cons :: forall (m :: Type -> Type). Applicative m => Branch0 m -> Branch m -> Branch m

mergeNode :: Applicative m => Branch0 m -> (CausalHash, m (Branch m)) -> (CausalHash, m (Branch m)) -> Branch m

Construct a two-parent merge node.

uncons :: Applicative m => Branch m -> m (Maybe (Branch0 m, Branch m))

empty :: forall (m :: Type -> Type). Branch m

empty0 :: forall (m :: Type -> Type). Branch0 m

discardHistory :: forall (m :: Type -> Type). Applicative m => Branch m -> Branch m

Discards the history of a Branch and its children, recursively

discardHistory0 :: forall (m :: Type -> Type). Applicative m => Branch0 m -> Branch0 m

Discards the history of a Branch0's children, recursively

transform :: Functor m => (forall a. m a -> n a) -> Branch m -> Branch n

transform0 :: Functor m => (forall a. m a -> n a) -> Branch0 m -> Branch0 n

Branch tests

isEmpty :: forall (m :: Type -> Type). Branch m -> Bool

Checks whether a branch is empty AND has no history.

isEmpty0 :: forall (m :: Type -> Type). Branch0 m -> Bool

isOne :: forall (m :: Type -> Type). Branch m -> Bool

before :: Monad m => Branch m -> Branch m -> m Bool

`before b1 b2` is true if b2 incorporates all of b1

lca :: Monad m => Branch m -> Branch m -> m (Maybe (Branch m))

properties

history_ :: forall (m :: Type -> Type) p f. (Profunctor p, Functor f) => p (UnwrappedBranch m) (f (UnwrappedBranch m)) -> p (Branch m) (f (Branch m))

head :: forall (m :: Type -> Type). Branch m -> Branch0 m

head_ :: forall (m :: Type -> Type) f. Functor f => (Branch0 m -> f (Branch0 m)) -> Branch m -> f (Branch m)

Update the head of the current causal. This re-hashes the current causal head after modifications.

headHash :: forall (m :: Type -> Type). Branch m -> CausalHash

namespaceHash :: forall (m :: Type -> Type). Branch m -> BranchHash

children_ :: forall (m :: Type -> Type) f. Functor f => (Map NameSegment (Branch m) -> f (Map NameSegment (Branch m))) -> Branch0 m -> f (Branch0 m)

nonEmptyChildren :: forall (m :: Type -> Type). Branch0 m -> Map NameSegment (Branch m)

namespaceStats :: forall (m :: Type -> Type). Branch0 m -> NamespaceStats

step

step :: forall (m :: Type -> Type). Applicative m => (Branch0 m -> Branch0 m) -> Branch m -> Branch m

Perform an update over the current branch and create a new causal step.

stepManyAt :: forall (m :: Type -> Type) f. (Monad m, Foldable f) => f (Path, Branch0 m -> Branch0 m) -> Branch m -> Branch m

Run a series of updates at specific locations, aggregating all changes into a single causal step. History is managed according to UpdateStrategy.

stepManyAtM :: forall (m :: Type -> Type) n f. (Monad m, Monad n, Foldable f) => f (Path, Branch0 m -> n (Branch0 m)) -> Branch m -> n (Branch m)

Run a series of updates at specific locations. History is managed according to the UpdateStrategy

stepEverywhere :: forall (m :: Type -> Type). Applicative m => (Branch0 m -> Branch0 m) -> Branch0 m -> Branch0 m

batchUpdates :: forall f (m :: Type -> Type). (Monad m, Foldable f) => f (Path, Branch0 m -> Branch0 m) -> Branch0 m -> Branch0 m

Perform updates over many locations within a branch by batching up operations on sub-branches as much as possible without affecting semantics. This operation does not create any causal conses, the operations are performed directly on the current head of the provided branch and child branches. It's the caller's responsibility to apply updates in history however they choose.

batchUpdatesM :: forall (m :: Type -> Type) n f. (Monad m, Monad n, Foldable f) => f (Path, Branch0 m -> n (Branch0 m)) -> Branch0 m -> n (Branch0 m)

Batch many updates. This allows us to apply the updates while minimizing redundant traversals. Semantics of operations are preserved by ensuring that all updates will always see changes by updates before them in the list.

This method does not step any branches on its own, all causal changes must be performed in the updates themselves, or this batch update must be provided to 'stepManyAt(M)'.

data UpdateStrategy

Constructors

CompressHistory
Compress all changes into a single causal cons. The resulting branch will have at most one new causal cons at each branch.

Note that this does NOT allow updates to add histories at children. E.g. if the root.editme branch has history: A -> B -> C and you use makeSetBranch to update it to a new branch with history X -> Y -> Z, CompressHistory will result in a history for root.editme of: A -> B -> C -> Z. A snapshot of the most recent state of the updated branch is appended to the existing history, if the new state is equal to the existing state, no new history nodes are appended.

AllowRewritingHistory
Preserves any history changes made within the update.

Note that this allows you to clobber the history child branches if you want. E.g. if the root.editme branch has history: A -> B -> C and you use makeSetBranch to update it to a new branch with history X -> Y -> Z, AllowRewritingHistory will result in a history for root.editme of: X -> Y -> Z. The history of the updated branch is replaced entirely.

addTermName :: forall (m :: Type -> Type). Referent -> NameSegment -> Branch0 m -> Branch0 m

addTypeName :: forall (m :: Type -> Type). TypeReference -> NameSegment -> Branch0 m -> Branch0 m

deleteTermName :: forall (m :: Type -> Type). Referent -> NameSegment -> Branch0 m -> Branch0 m

annihilateTermName :: forall (m :: Type -> Type). NameSegment -> Branch0 m -> Branch0 m

annihilateTypeName :: forall (m :: Type -> Type). NameSegment -> Branch0 m -> Branch0 m

deleteTypeName :: forall (m :: Type -> Type). TypeReference -> NameSegment -> Branch0 m -> Branch0 m

setChildBranch :: forall (m :: Type -> Type). NameSegment -> Branch m -> Branch0 m -> Branch0 m

Children queries

getAt :: forall (m :: Type -> Type). Path -> Branch m -> Maybe (Branch m)

getAt' :: forall (m :: Type -> Type). Path -> Branch m -> Branch m

getAt0 :: forall (m :: Type -> Type). Path -> Branch0 m -> Branch0 m

modifyAt :: forall (m :: Type -> Type). Applicative m => Path -> (Branch m -> Branch m) -> Branch m -> Branch m

modifyAtM :: forall n (m :: Type -> Type). (Functor n, Applicative m) => Path -> (Branch m -> n (Branch m)) -> Branch m -> n (Branch m)

children0 :: forall (m :: Type -> Type) p f. (Indexable NameSegment p, Applicative f) => p (Branch0 m) (f (Branch0 m)) -> Branch0 m -> f (Branch0 m)

Traverse the head branch of all direct children. The index of the traversal is the name of that child branch according to the parent.

Libdep manipulations

libdeps_ :: forall (m :: Type -> Type) f. Applicative f => (Map NameSegment (Branch m) -> f (Map NameSegment (Branch m))) -> Branch0 m -> f (Branch0 m)

hasDefnsInLib :: forall (m :: Type -> Type). Branch0 m -> Bool

Are there any pesky definitions (terms or types) directly in lib.*, where they shouldn't be?

withoutLib :: forall (m :: Type -> Type). Branch0 m -> Branch0 m

Remove any lib subtrees reachable within the branch. Note: This DOES affect the hash.

withoutTransitiveLibs :: forall (m :: Type -> Type). Branch0 m -> Branch0 m

Remove any transitive libs reachable within the branch. Note: This DOES affect the hash.

deleteLibdep :: forall (m :: Type -> Type). NameSegment -> Branch0 m -> Branch0 m

deleteLibdep name branch deletes the libdep named name from branch, if it exists.

deleteLibdeps :: forall (m :: Type -> Type). Branch0 m -> Branch0 m

deleteLibdeps branch deletes all libdeps from branch.

setLibdeps :: forall (m :: Type -> Type). Branch0 m -> Branch0 m -> Branch0 m

setLibdeps libdeps branch sets branch's libdeps to libdeps.

Branch termstypesedits

Termtypeedits lenses

terms_ :: forall (m :: Type -> Type) f. Functor f => (Star Referent NameSegment -> f (Star Referent NameSegment)) -> Branch0 m -> f (Branch0 m)

types_ :: forall (m :: Type -> Type) f. Functor f => (Star TypeReference NameSegment -> f (Star TypeReference NameSegment)) -> Branch0 m -> f (Branch0 m)

edits_ :: forall m f. Functor f => (Map NameSegment (PatchHash, m Patch) -> f (Map NameSegment (PatchHash, m Patch))) -> Branch0 m -> f (Branch0 m)

Term/type queries

deepDefns :: forall (m :: Type -> Type). Branch0 m -> Defns (Relation Referent Name) (Relation TypeReference Name)

deepDefnsRefs :: forall (m :: Type -> Type). Branch0 m -> DefnsF Set TermReference TypeReference

deepDefnsIds :: forall (m :: Type -> Type). Branch0 m -> DefnsF Set TermReferenceId TypeReferenceId

deepTerms :: forall (m :: Type -> Type). Branch0 m -> Relation Referent Name

deepTypes :: forall (m :: Type -> Type). Branch0 m -> Relation TypeReference Name

deepPaths :: forall (m :: Type -> Type). Branch0 m -> Set Path

deepReferents :: forall (m :: Type -> Type). Branch0 m -> Set Referent

deepTermReferences :: forall (m :: Type -> Type). Branch0 m -> Set TermReference

deepTermReferenceIds :: forall (m :: Type -> Type). Branch0 m -> Set TermReferenceId

deepTypeReferences :: forall (m :: Type -> Type). Branch0 m -> Set TypeReference

deepTypeReferenceIds :: forall (m :: Type -> Type). Branch0 m -> Set TypeReferenceId

asUnconflicted :: Branch0 m -> Either (Defn (Conflicted Name Referent) (Conflicted Name TypeReference)) UnconflictedLocalDefnsView

consBranchSnapshot :: forall (m :: Type -> Type). Monad m => Branch m -> Branch m -> Branch m

head consBranchSnapshot base Cons's the current state of head onto base as-is. Consider whether you really want this behaviour or the behaviour of squashMerge That is, it does not perform any common ancestor detection, or change reconciliation, it sets the current state of the base branch to the new state as a new causal step (or returns the existing base if there are no)

Orphan instances

AsEmpty (Branch m)

Instance details

Methods

_Empty :: Prism' (Branch m) ()

ContentAddressable (Branch0 m)

Instance details

Methods

contentHash :: Branch0 m -> Hash