Unison.Codebase.BranchUtil

Branch creation

fromNames :: Monad m => Names -> Branch m

Creates a branch containing all of the given names, with a single history node.

Branch queries

getBranch :: Split -> Branch0 m -> Maybe (Branch m)

getTerm :: HQSplit -> Branch0 m -> Set Referent

getType :: HQSplit -> Branch0 m -> Set Reference

getTermMetadataAt :: (Path, a) -> Referent -> Branch0 m -> Metadata

getTypeMetadataAt :: (Path, a) -> Reference -> Branch0 m -> Metadata

getTermMetadataHQNamed :: (Path, HQSegment) -> Branch0 m -> R4 Referent NameSegment

getTypeMetadataHQNamed :: (Path, HQSegment) -> Branch0 m -> R4 Reference NameSegment

Branch modifications

makeSetBranch :: Split -> Branch m -> (Path, Branch0 m -> Branch0 m)

makeDeleteBranch :: Applicative m => Split -> (Path, Branch0 m -> Branch0 m)

"delete"s a branch by cons'ing an empty Branch0 onto the history at that location. See also makeObliterateBranch.

makeObliterateBranch :: Split -> (Path, Branch0 m -> Branch0 m)

Erase a branch and its history See also makeDeleteBranch. Note that this requires a AllowRewritingHistory update strategy to behave correctly.

makeAddTypeName :: Split -> Reference -> Metadata -> (Path, Branch0 m -> Branch0 m)

makeDeleteTypeName :: Split -> Reference -> (Path, Branch0 m -> Branch0 m)

makeAddTermName :: Split -> Referent -> Metadata -> (Path, Branch0 m -> Branch0 m)

makeDeleteTermName :: Split -> Referent -> (Path, Branch0 m -> Branch0 m)

makeDeletePatch :: Split -> (Path, Branch0 m -> Branch0 m)

makeReplacePatch :: Applicative m => Split -> Patch -> (Path, Branch0 m -> Branch0 m)