Unison.Codebase.Editor.Git

Documentation

gitIn :: MonadIO m => GitRepo -> [Text] -> m ()

Run a git command in the repository at localPath

gitTextIn :: MonadIO m => GitRepo -> [Text] -> m Text

Run a git command in the repository at localPath and capture stdout

gitInCaptured :: MonadIO m => GitRepo -> [Text] -> m (Bool, Text, Text)

like gitIn, but silences all output from the command and returns whether the command succeeded.

withRepo :: forall m a. MonadUnliftIO m => ReadGitRepo -> GitBranchBehavior -> (GitRepo -> m a) -> m (Either GitProtocolError a)

Clone or fetch an updated copy of the provided repository and check out the expected ref, then provide the action with a path to the codebase in that repository. Note that the repository provided to the action is temporary, it will be removed when the action completes or fails.

withIOError :: MonadIO m => IO a -> (IOException -> m a) -> m a

Perform an IO action, passing any IO exception to handler

withStatus :: MonadIO m => String -> m a -> m a

withIsolatedRepo :: forall m r. MonadUnliftIO m => GitRepo -> Text -> Maybe Text -> (GitRepo -> m r) -> m (Either GitProtocolError r)

Run an action on an isolated copy of the provided repo. The repo is deleted when the action exits or fails. A branch or tag to check out from the source repo may be specified.

debugGit :: Bool

gitDirToPath :: GitRepo -> FilePath

gitVerbosity :: [Text]

data GitBranchBehavior

Define what to do if the repo we're pulling/pushing doesn't have the specified branch.

Constructors

CreateBranchIfMissing

RequireExistingBranch

data GitRepo

A path to a git repository.

Constructors

Bare FilePath

Worktree FilePath

Exported for testing

gitCacheDir :: MonadIO m => Text -> m FilePath