Unison.Codebase.Editor.Output

Documentation

data Output v

Constructors

Success

NoUnisonFile

PrintMessage (Pretty ColorText)

InvalidSourceName String

SourceLoadFailed String

NoMainFunction String PrettyPrintEnv [Type v Ann]

BadMainFunction String (Type v Ann) PrettyPrintEnv [Type v Ann]

BranchEmpty (Either ShortBranchHash Path')

BranchNotEmpty Path'

LoadPullRequest ReadRemoteNamespace ReadRemoteNamespace Path' Path' Path' Path'

CreatedNewBranch Absolute

BranchAlreadyExists Path'

PatchAlreadyExists Split'

NoExactTypeMatches

TypeAlreadyExists Split' (Set Reference)

TypeParseError String (Err v)

ParseResolutionFailures String [ResolutionFailure v Ann]

TypeHasFreeVars (Type v Ann)

TermAlreadyExists Split' (Set Referent)

LabeledReferenceAmbiguous Int (HashQualified Name) (Set LabeledDependency)

LabeledReferenceNotFound (HashQualified Name)

DeleteNameAmbiguous Int HQSplit' (Set Referent) (Set Reference)

TermAmbiguous (HashQualified Name) (Set Referent)

HashAmbiguous ShortHash (Set Referent)

BranchHashAmbiguous ShortBranchHash (Set ShortBranchHash)

BranchNotFound Path'

NameNotFound HQSplit'

PatchNotFound Split'

TypeNotFound HQSplit'

TermNotFound HQSplit'

TypeNotFound' ShortHash

TermNotFound' ShortHash

TypeTermMismatch (HashQualified Name) (HashQualified Name)

SearchTermsNotFound [HashQualified Name]

DeleteBranchConfirmation [(Path', (Names, [SearchResult' v Ann]))]

CantDelete PrettyPrintEnv [SearchResult' v Ann] [SearchResult' v Ann]

DeleteEverythingConfirmation

DeletedEverything

ListNames Int [(Reference, Set (HashQualified Name))] [(Referent, Set (HashQualified Name))]

ListOfDefinitions PrettyPrintEnv ListDetailed [SearchResult' v Ann]

ListOfLinks PrettyPrintEnv [(HashQualified Name, Reference, Maybe (Type v Ann))]

ListShallow PrettyPrintEnv [ShallowListEntry v Ann]

ListOfPatches (Set Name)

SlurpOutput Input PrettyPrintEnv (SlurpResult v)

ParseErrors Text [Err v]

TypeErrors Text PrettyPrintEnv [ErrorNote v Ann]

CompilerBugs Text PrettyPrintEnv [CompilerBug v Ann]

DisplayConflicts (Relation Name Referent) (Relation Name Reference)

EvaluationFailure Error

Evaluated SourceFileContents PrettyPrintEnv [(v, Term v ())] (Map v (Ann, WatchKind, Term v (), IsCacheHit))

Typechecked SourceName PrettyPrintEnv (SlurpResult v) (TypecheckedUnisonFile v Ann)

DisplayRendered (Maybe FilePath) (Pretty ColorText)

DisplayDefinitions (Maybe FilePath) PrettyPrintEnvDecl (Map Reference (DisplayObject () (Decl v Ann))) (Map Reference (DisplayObject (Type v Ann) (Term v Ann)))

TodoOutput PrettyPrintEnvDecl (TodoOutput v Ann)
Invariant: there's at least one conflict or edit in the TodoOutput.

TestIncrementalOutputStart PrettyPrintEnv (Int, Int) Reference (Term v Ann)

TestIncrementalOutputEnd PrettyPrintEnv (Int, Int) Reference (Term v Ann)

TestResults TestReportStats PrettyPrintEnv ShowSuccesses ShowFailures [(Reference, Text)] [(Reference, Text)]

CantUndo UndoFailureReason

ListEdits Patch PrettyPrintEnv

BustedBuiltins (Set Reference) (Set Reference)

GitError Input GitError

ConfiguredMetadataParseError Path' String (Pretty ColorText)

NoConfiguredGitUrl PushPull Path'

ConfiguredGitUrlParseError PushPull Path' Text String

DisplayLinks PrettyPrintEnvDecl Metadata (Map Reference (DisplayObject () (Decl v Ann))) (Map Reference (DisplayObject (Type v Ann) (Term v Ann)))

MetadataMissingType PrettyPrintEnv Referent

TermMissingType Reference

MetadataAmbiguous (HashQualified Name) PrettyPrintEnv [Referent]

NothingToPatch PatchPath Path'

PatchNeedsToBeConflictFree

PatchInvolvesExternalDependents PrettyPrintEnv (Set Reference)

WarnIncomingRootBranch ShortBranchHash (Set ShortBranchHash)

StartOfCurrentPathHistory

History (Maybe Int) [(ShortBranchHash, Diff)] HistoryTail

ShowReflog [ReflogEntry]

PullAlreadyUpToDate ReadRemoteNamespace Path'

MergeAlreadyUpToDate Path' Path'

PreviewMergeAlreadyUpToDate Path' Path'

NoConflictsOrEdits
No conflicts or edits remain for the current patch.

NotImplemented

NoBranchWithHash ShortBranchHash

ListDependencies Int LabeledDependency [(Name, Reference)] (Set Reference)

ListDependents Int LabeledDependency [(Name, Reference)] (Set Reference)

DumpNumberedArgs NumberedArgs

DumpBitBooster Hash (Map Hash [Hash])

DumpUnisonFileHashes Int [(Name, Id)] [(Name, Id)] [(Name, Id)]

BadName String

DefaultMetadataNotification

BadRootBranch GetRootBranchError

CouldntLoadBranch Hash

NoOp

data NumberedOutput v

Constructors

ShowDiffNamespace Absolute Absolute PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterUndo PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterDeleteDefinitions PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterDeleteBranch Absolute PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterModifyBranch Path' Absolute PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterMerge Path' Absolute PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterMergePropagate Path' Absolute Path' PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterMergePreview Path' Absolute PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterPull Path' Absolute PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterCreatePR ReadRemoteNamespace ReadRemoteNamespace PrettyPrintEnv (BranchDiffOutput v Ann)

ShowDiffAfterCreateAuthor NameSegment Path' Absolute PrettyPrintEnv (BranchDiffOutput v Ann)

type NumberedArgs = [String]

type ListDetailed = Bool

data HistoryTail

Constructors

EndOfLog ShortBranchHash

MergeTail ShortBranchHash [ShortBranchHash]

PageEnd ShortBranchHash Int

data TestReportStats

Constructors

CachedTests TotalCount CachedCount

NewlyComputed

data UndoFailureReason

Constructors

CantUndoPastStart

CantUndoPastMerge

data PushPull

Constructors

Push

Pull

data ReflogEntry

Constructors

ReflogEntry

Fields

hash :: ShortBranchHash

reason :: Text

pushPull :: a -> a -> PushPull -> a

isFailure :: Ord v => Output v -> Bool

isNumberedFailure :: NumberedOutput v -> Bool