Unison.Codebase.Editor.RemoteRepo

Documentation

data ReadRepo

Constructors

ReadRepoGit ReadGitRepo

ReadRepoShare ShareCodeserver

data ShareCodeserver

Constructors

DefaultCodeserver

CustomCodeserver CodeserverURI

newtype ShareUserHandle

Constructors

ShareUserHandle

Fields

shareUserHandleToText :: Text

displayShareCodeserver :: ShareCodeserver -> ShareUserHandle -> Path -> Text

>>> :set -XOverloadedLists
>>> import Data.Maybe (fromJust)
>>> import Network.URI
>>> displayShareCodeserver DefaultCodeserver "share" ["base", "List"]
"share.base.List"
>>> displayShareCodeserver DefaultCodeserver "share" []
"share"
>>> displayShareCodeserver (CustomCodeserver . fromJust $ parseURI "https://share-next.unison-lang.org/api" >>= codeserverFromURI ) "unison" ["base", "List"]
"share(https://share-next.unison-lang.org:443/api).unison.base.List"

data ReadGitRepo

Constructors

ReadGitRepo

Fields

url :: Text

ref :: Maybe Text

data WriteRepo

Constructors

WriteRepoGit WriteGitRepo

WriteRepoShare ShareCodeserver

data WriteGitRepo

Constructors

WriteGitRepo

Fields

url :: Text

branch :: Maybe Text

writeToRead :: WriteRepo -> ReadRepo

writeToReadGit :: WriteGitRepo -> ReadGitRepo

writeNamespaceToRead :: WriteRemoteNamespace Void -> ReadRemoteNamespace void

printReadGitRepo :: ReadGitRepo -> Text

printWriteGitRepo :: WriteGitRepo -> Text

printReadRemoteNamespace :: (a -> Text) -> ReadRemoteNamespace a -> Text

print remote namespace

printWriteRemoteNamespace :: WriteRemoteNamespace (ProjectAndBranch ProjectName ProjectBranchName) -> Text

Render a WriteRemoteNamespace as text.

maybePrintPath :: Path -> Text

data ReadRemoteNamespace a

Constructors

ReadRemoteNamespaceGit !ReadGitRemoteNamespace

ReadShare'LooseCode !ReadShareLooseCode

ReadShare'ProjectBranch !a
A remote project+branch, specified by name (e.g. @unisonbasemain). Currently assumed to be hosted on Share, though we could include a ShareCodeserver in here, too.

data ReadGitRemoteNamespace

Constructors

ReadGitRemoteNamespace

Fields

repo :: !ReadGitRepo

sch :: !(Maybe ShortCausalHash)

path :: !Path

data ReadShareLooseCode

Constructors

ReadShareLooseCode

Fields

server :: !ShareCodeserver

repo :: !ShareUserHandle

path :: !Path

isPublic :: ReadShareLooseCode -> Bool

data WriteRemoteNamespace a

Constructors

WriteRemoteNamespaceGit !WriteGitRemoteNamespace

WriteRemoteNamespaceShare !WriteShareRemoteNamespace

WriteRemoteProjectBranch a

remotePath_ :: Lens' (WriteRemoteNamespace Void) Path

A lens which focuses the path of a remote namespace.

data WriteGitRemoteNamespace

Constructors

WriteGitRemoteNamespace

Fields

repo :: !WriteGitRepo

path :: !Path

data WriteShareRemoteNamespace

Constructors

WriteShareRemoteNamespace

Fields

server :: !ShareCodeserver

repo :: !ShareUserHandle

path :: !Path