Unison.Codebase.Init

Documentation

data Init m v a

Constructors

Init

Fields

openCodebase :: DebugName -> CodebasePath -> m (Either Pretty (m (), Codebase m v a))

open an existing codebase

createCodebase' :: DebugName -> CodebasePath -> m (Either CreateCodebaseError (m (), Codebase m v a))

create a new codebase

codebasePath :: CodebasePath -> CodebasePath

given a codebase root, and given that the codebase root may have other junk in it, give the path to the "actual" files; e.g. what a forked transcript should clone.

type DebugName = String

data InitError

Constructors

NoCodebaseFoundAtSpecifiedDir

FoundV1Codebase

CouldntCreateCodebase Pretty

data CodebaseInitOptions

Constructors

Home CodebasePath

Specified SpecifiedCodebase

data InitResult m v a

Constructors

OpenedCodebase CodebasePath (FinalizerAndCodebase m v a)

CreatedCodebase CodebasePath (FinalizerAndCodebase m v a)

Error CodebasePath InitError

data SpecifiedCodebase

Constructors

CreateWhenMissing CodebasePath

DontCreateWhenMissing CodebasePath

type Pretty = Pretty ColorText

createCodebase :: MonadIO m => Init m v a -> DebugName -> CodebasePath -> m (Either Pretty (m (), Codebase m v a))

initCodebaseAndExit :: MonadIO m => Init m Symbol Ann -> DebugName -> Maybe CodebasePath -> m ()

try to init a codebase where none exists and then exit regardless (i.e. `ucm -codebase dir init`)

openOrCreateCodebase :: MonadIO m => Init m v a -> DebugName -> CodebaseInitOptions -> m (InitResult m v a)

openNewUcmCodebaseOrExit :: MonadIO m => Init m Symbol Ann -> DebugName -> CodebasePath -> m (m (), Codebase m Symbol Ann)