Unison.Codebase.Init.Type

Documentation

data Init m v a

Constructors

Init

Fields

openCodebase :: DebugName -> CodebasePath -> m (Either Pretty (m (), Codebase m v a))

open an existing codebase

createCodebase' :: DebugName -> CodebasePath -> m (Either CreateCodebaseError (m (), Codebase m v a))

create a new codebase

codebasePath :: CodebasePath -> CodebasePath

given a codebase root, and given that the codebase root may have other junk in it, give the path to the "actual" files; e.g. what a forked transcript should clone.