Unison.Codebase.IntegrityCheck

Description

There are many invariants we expect to hold in our sqlite database and on codebase objects which we can't maintain using database checks. This module performs checks for some of these invariants, which can be useful to run after performing potentially dangerous operations like migrations.

Documentation

integrityCheckFullCodebase :: Transaction IntegrityResult

Performs all available integrity checks.

integrityCheckAllBranches :: Transaction IntegrityResult

Performs a bevy of checks on branch objects and their relation to causals.

integrityCheckAllCausals :: Transaction IntegrityResult

Performs a bevy of checks on causals.

prettyPrintIntegrityErrors :: Foldable f => f IntegrityError -> Pretty ColorText

data IntegrityResult

Constructors

IntegrityErrorDetected (NESet IntegrityError)

NoIntegrityErrors