Unison.Codebase.Path

Documentation

newtype Path

Constructors

Path

Fields

toSeq :: Seq NameSegment

newtype Path'

Constructors

Path'

Fields

unPath' :: Either Absolute Relative

newtype Absolute

Constructors

Absolute

Fields

unabsolute :: Path

pattern AbsolutePath' :: Absolute -> Path'

newtype Relative

Constructors

Relative

Fields

unrelative :: Path

pattern RelativePath' :: Relative -> Path'

class Resolve l r o where

Methods

resolve :: l -> r -> o

pattern Empty :: Path

pattern (:<) :: Cons b b a a => a -> b -> b infixr 5

pattern (:>) :: Snoc a a b b => a -> b -> a infixl 5

singleton :: NameSegment -> Path

uncons :: Path -> Maybe (NameSegment, Path)

empty :: Path

isAbsolute :: Path' -> Bool

isRelative :: Path' -> Bool

absoluteEmpty :: Absolute

absoluteEmpty' :: Path'

relativeEmpty :: Relative

relativeEmpty' :: Path'

currentPath :: Path'

prefix :: Absolute -> Path' -> Path

unprefix :: Absolute -> Path' -> Path

examples: unprefix .foo.bar .blah == .blah (absolute paths left alone) unprefix .foo.bar id == id (relative paths starting w/ nonmatching prefix left alone) unprefix .foo.bar foo.bar.baz == baz (relative paths w/ common prefix get stripped)

prefixName :: Absolute -> Name -> Name

unprefixName :: Absolute -> Name -> Maybe Name

Remove a path prefix from a name. Returns Nothing if there are no remaining segments to construct the name from.

>>> unprefixName (Absolute $ fromList ["base", "List"]) (Name.unsafeFromText "base.List.map")
Just (Name Relative (NameSegment {toText = "map"} :| []))

type HQSplit = (Path, HQSegment)

type Split = (Path, NameSegment)

type Split' = (Path', NameSegment)

type HQSplit' = (Path', HQSegment)

ancestors :: Absolute -> Seq Absolute

utilities

longestPathPrefix :: Path -> Path -> (Path, Path, Path)

Finds the longest shared path prefix of two paths. Returns (shared prefix, path to first location from shared prefix, path to second location from shared prefix)

>>> longestPathPrefix ("a" :< "b" :< "x" :< Empty) ("a" :< "b" :< "c" :< Empty)
(a.b,x,c)

>>> longestPathPrefix Empty ("a" :< "b" :< "c" :< Empty)
(,,a.b.c)

tests

isCurrentPath :: Path' -> Bool

isRoot :: Absolute -> Bool

isRoot' :: Path' -> Bool

things that could be replaced with Convert instances

absoluteToPath' :: Absolute -> Path'

fromList :: [NameSegment] -> Path

fromName :: Name -> Path

fromName' :: Name -> Path'

fromPath' :: Path' -> Path

unsafeParseText :: Text -> Path

unsafeParseText' :: Text -> Path'

Construct a Path' from a text

>>> fromText' "a.b.c"
a.b.c

>>> fromText' ".a.b.c"
.a.b.c

>>> show $ fromText' ""
""

toAbsoluteSplit :: Absolute -> (Path', a) -> (Absolute, a)

toSplit' :: Path' -> Maybe (Path', NameSegment)

toList :: Path -> [NameSegment]

toName :: Path -> Maybe Name

toName' :: Path' -> Maybe Name

Convert a Path' to a Name

unsafeToName :: Path -> Name

unsafeToName' :: Path' -> Name

Convert a Path' to a Name

toText :: Path -> Text

Note: This treats the path as relative.

toText' :: Path' -> Text

unsplit :: Split -> Path

unsplit' :: Split' -> Path'

unsplitAbsolute :: (Absolute, NameSegment) -> Absolute

unsplitHQ :: HQSplit -> HashQualified Path

unsplitHQ' :: HQSplit' -> HashQualified Path'

things that could be replaced with Parse instances

splitFromName :: Name -> Split

>>> splitFromName "a.b.c"
(a.b,c)

>>> splitFromName "foo"
(,foo)

splitFromName' :: Name -> Split'

hqSplitFromName' :: Name -> Maybe HQSplit'

things that could be replaced with Cons instances

cons :: NameSegment -> Path -> Path

things that could be replaced with Snoc instances

snoc :: Path -> NameSegment -> Path

unsnoc :: Path -> Maybe (Path, NameSegment)

class Convert a b where

Methods

convert :: a -> b

Orphan instances

Convert (path, NameSegment) (path, HQSegment)

Instance details

Methods

convert :: (path, NameSegment) -> (path, HQSegment)

Convert path0 path1 => Convert (path0, name) (path1, name)

Instance details

Methods

convert :: (path0, name) -> (path1, name)