Unison.Codebase.ProjectPath

Documentation

data ProjectPathG proj branch

Constructors

ProjectPath

Fields

project :: proj

branch :: branch

absPath :: Absolute

type ProjectPathIds = ProjectPathG ProjectId ProjectBranchId

type ProjectPathNames = ProjectPathG ProjectName ProjectBranchName

type ProjectPath = ProjectPathG Project ProjectBranch

fromProjectAndBranch :: ProjectAndBranch Project ProjectBranch -> Absolute -> ProjectPath

projectBranchRoot :: ProjectAndBranch Project ProjectBranch -> ProjectPath

toRoot :: ProjectPath -> ProjectPath

Discard any path within the project and get the project's root

absPath_ :: Lens' (ProjectPathG p b) Absolute

path_ :: Lens' (ProjectPathG p b) Path

path :: ProjectPathG p b -> Path

toProjectAndBranch :: ProjectPathG p b -> ProjectAndBranch p b

projectAndBranch_ :: Lens (ProjectPathG p b) (ProjectPathG p' b') (ProjectAndBranch p b) (ProjectAndBranch p' b')

toText :: ProjectPathG Project ProjectBranch -> Text

toIds :: ProjectPath -> ProjectPathIds

Project a project context into a project path of just IDs

toNames :: ProjectPath -> ProjectPathNames

Project a project context into a project path of just names

projectPathParser :: Parser ProjectPathNames

parseProjectPath :: Text -> Either Text ProjectPathNames

Re-exports, this also helps with using dot-notation

data ProjectAndBranch a b

A generic data structure that contains information about a project and a branch in that project.

Constructors

ProjectAndBranch

Fields

project :: a

branch :: b

data Project

A project.

Constructors

Project

Fields

projectId :: !ProjectId

name :: !ProjectName

data ProjectBranch

A project branch.

Constructors

ProjectBranch

Fields

projectId :: !ProjectId

branchId :: !ProjectBranchId

name :: !ProjectBranchName

parentBranchId :: !(Maybe ProjectBranchId)