Unison.Codebase.PushBehavior

Description

This module defines the PushBehavior type.

Documentation

data PushBehavior

How a push behaves.

Constructors

ForcePush

RequireEmpty
The namespace being pushed to is required to be empty.

RequireNonEmpty
The namespace being pushed to is required to be non-empty