Unison.Codebase.RootBranchCache

Documentation

data RootBranchCache

newEmptyRootBranchCache :: STM RootBranchCache

newEmptyRootBranchCacheIO :: MonadIO m => m RootBranchCache

fetchRootBranch :: forall m. MonadUnliftIO m => RootBranchCache -> m (Branch Transaction) -> m (Branch Transaction)

withLock

Arguments

:: forall m r. MonadUnliftIO m

=> RootBranchCache

-> ((forall x. m x -> m x) -> Maybe (Branch Transaction) -> m r)
Perform an action with the cached value

-> (r -> Maybe (Branch Transaction))
compute value to restore to the cache

-> m r

Take a cache lock so that no other thread can read or write to the cache, perform an action with the cached value, then restore the cache to Empty or Full