Unison.Codebase.ShortCausalHash

Documentation

toString :: ShortCausalHash -> String

toHash :: Coercible Hash h => ShortCausalHash -> Maybe h

fromHash :: Int -> CausalHash -> ShortCausalHash

fromFullHash :: Coercible h Hash => h -> ShortCausalHash

This allows a full hash to be preserved as a ShortCausalHash.

ShortCausalHash is used for input when we expect a user to enter a hash on the command line, so they aren’t required to enter the full hash. However, these inputs may also come from an internal source, and in such cases, there is no reason to truncate the hash.

fromText :: Text -> Maybe ShortCausalHash

newtype ShortCausalHash

Causal Hash Prefix

Constructors

ShortCausalHash

Fields

toText :: Text