Unison.Codebase.SqliteCodebase

Documentation

init :: (HasCallStack, MonadUnliftIO m) => Init m Symbol Ann

initWithSetup :: (HasCallStack, MonadUnliftIO m) => Transaction () -> Init m Symbol Ann

Like init, but allows passing in an action to be perform when a new codebase is created.

data MigrationStrategy

Constructors

MigrateAutomatically BackupStrategy VacuumStrategy
Perform a migration immediately if one is required.

MigrateAfterPrompt BackupStrategy VacuumStrategy
Prompt the user that a migration is about to occur, continue after acknownledgment

DontMigrate
Triggers an OpenCodebaseRequiresMigration error instead of migrating

data BackupStrategy

Constructors

Backup

NoBackup

data VacuumStrategy

Constructors

Vacuum

NoVacuum

data CodebaseLockOption

Constructors

DoLock

DontLock

copyCodebase :: MonadIO m => CodebasePath -> CodebasePath -> m ()

Given two codebase roots (e.g. "./mycodebase"), safely copy the codebase at the source to the destination. Note: this does not copy the .unisonConfig file.