Unison.Codebase.SqliteCodebase.Conversions

Documentation

sch1to2 :: ShortCausalHash -> ShortCausalHash

decltype2to1 :: DeclType -> ConstructorType

decltype1to2 :: ConstructorType -> DeclType

watchKind1to2 :: WatchKind -> WatchKind

watchKind2to1 :: WatchKind -> WatchKind

term1to2 :: Hash -> Term Symbol Ann -> Term Symbol

term2to1 :: forall m. Monad m => Hash -> (Reference -> m ConstructorType) -> Term Symbol -> m (Term Symbol Ann)

termComponent1to2 :: Hash -> [(Term Symbol Ann, Type Symbol a)] -> [(Term Symbol, TypeT Symbol)]

decl2to1 :: Hash -> Decl Symbol -> Decl Symbol Ann

decl1to2 :: Hash -> Decl Symbol a -> Decl Symbol

symbol2to1 :: Symbol -> Symbol

symbol1to2 :: Symbol -> Symbol

rreference2to1 :: Hash -> Reference' Text (Maybe Hash) -> Reference

rreference1to2 :: Hash -> Reference -> Reference' Text (Maybe Hash)

rreferenceid2to1 :: Hash -> Id' (Maybe Hash) -> Id

rreferenceid1to2 :: Hash -> Id -> Id' (Maybe Hash)

branchHash1to2 :: NamespaceHash m -> BranchHash

branchHash2to1 :: forall m. BranchHash -> NamespaceHash m

reference2to1 :: Reference -> Reference

reference1to2 :: Reference -> Reference

referenceid1to2 :: Id -> Id

referenceid2to1 :: Id -> Id

rreferent2to1 :: Applicative m => Hash -> (Reference -> m ConstructorType) -> ReferentH -> m Referent

rreferent1to2 :: Hash -> Referent -> ReferentH

referent2to1 :: Applicative m => (Reference -> m ConstructorType) -> Referent -> m Referent

referent2to1UsingCT :: ConstructorType -> Referent -> Referent

Like referent2to1, but uses the provided constructor type directly

referent1to2 :: Referent -> Referent

referentid1to2 :: Id -> Id

referentid2to1 :: Applicative m => (Reference -> m ConstructorType) -> Id -> m Id

constructorType1to2 :: ConstructorType -> ConstructorType

constructorType2to1 :: ConstructorType -> ConstructorType

ttype2to1 :: Type Symbol -> Type Symbol Ann

dtype2to1 :: Hash -> Type Symbol -> Type Symbol Ann

type2to1' :: (r -> Reference) -> TypeR r Symbol -> Type Symbol Ann

dtype1to2 :: Hash -> Type Symbol a -> TypeD Symbol

ttype1to2 :: Type Symbol a -> TypeT Symbol

type1to2' :: (Reference -> r) -> Type Symbol a -> TypeR r Symbol

causalbranch2to1 :: Monad m => BranchCache m -> (Reference -> m ConstructorType) -> CausalBranch m -> m (Branch m)

forces loading v1 branches even if they may not exist

causalbranch2to1' :: Monad m => BranchCache m -> (Reference -> m ConstructorType) -> CausalBranch m -> m (UnwrappedBranch m)

causalbranch1to2 :: forall m. Monad m => Branch m -> CausalBranch m

patch2to1 :: Patch -> Patch

patch1to2 :: Patch -> Patch

branch2to1 :: Monad m => BranchCache m -> (Reference -> m ConstructorType) -> Branch m -> m (Branch0 m)

referent2toshorthash1 :: Maybe Int -> Referent -> ShortHash

Generates a v1 short hash from a v2 referent. Also shortens the hash to the provided length. If Nothing, it will include the full length hash.

reference2toshorthash1 :: Maybe Int -> Reference -> ShortHash

Generates a v1 short hash from a v2 reference. Also shortens the hash to the provided length. If Nothing, it will include the full length hash.