Unison.Codebase.SqliteCodebase.Migrations

Documentation

migrations

Arguments

:: (Reference -> Transaction ConstructorType)
A getDeclType-like lookup, possibly backed by a cache.

-> TVar (Map Hash TermBufferEntry)

-> TVar (Map Hash DeclBufferEntry)

-> Map SchemaVersion (Transaction ())

Mapping from schema version to the migration required to get there. Each migration may only be run on a schema of its immediate predecessor, E.g. The migration at index 2 must be run on a codebase at version 1.

ensureCodebaseIsUpToDate

Arguments

:: MonadUnliftIO m

=> LocalOrRemote

-> CodebasePath

-> (Reference -> Transaction ConstructorType)
A getDeclType-like lookup, possibly backed by a cache.

-> TVar (Map Hash TermBufferEntry)

-> TVar (Map Hash DeclBufferEntry)

-> Connection

-> m (Either OpenCodebaseError ())

Migrates a codebase up to the most recent version known to ucm. This is a No-op if it's up to date Returns an error if the schema version is newer than this ucm knows about.

backupCodebase :: CodebasePath -> IO ()

Copy the sqlite database to a new file with a unique name based on current time.