Unison.Codebase.SqliteCodebase.Paths

Documentation

codebasePath :: FilePath

Prefer makeCodebasePath or makeCodebaseDirPath when possible.

makeCodebasePath :: CodebasePath -> FilePath

Makes a path to a sqlite database from a codebase path.

makeCodebaseDirPath :: CodebasePath -> FilePath

Makes a path to the location where sqlite files are stored within a codebase path.

backupCodebasePath :: SchemaVersion -> NominalDiffTime -> FilePath

Makes a path to store a backup of a sqlite database given the current time.

lockfilePath :: CodebasePath -> FilePath