Unison.Codebase.Type

Documentation

data Codebase m v a

Abstract interface to a user's codebase.

Constructors

Codebase

Fields

getTerm :: Id -> Transaction (Maybe (Term v a))

Get a user-defined term from the codebase.

Note that it is possible to call $sel:putTerm:Codebase, then $sel:getTerm:Codebase, and receive Nothing, per the semantics of $sel:putTerm:Codebase.

getTypeOfTermImpl :: Id -> Transaction (Maybe (Type v a))

Get the type of a user-defined term.

Note that it is possible to call $sel:putTerm:Codebase, then $sel:getTypeOfTermImpl:Codebase, and receive Nothing, per the semantics of $sel:putTerm:Codebase.

getTypeDeclaration :: Id -> Transaction (Maybe (Decl v a))

Get a type declaration.

Note that it is possible to call $sel:putTypeDeclaration:Codebase, then $sel:getTypeDeclaration:Codebase, and receive Nothing, per the semantics of $sel:putTypeDeclaration:Codebase.

getDeclType :: Reference -> Transaction ConstructorType

Get the type of a given decl.

putTerm :: Id -> Term v a -> Type v a -> Transaction ()

Enqueue the put of a user-defined term (with its type) into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the term's (and its type's) references are stored as well.

putTermComponent :: Hash -> [(Term v a, Type v a)] -> Transaction ()

putTypeDeclaration :: Id -> Decl v a -> Transaction ()

Enqueue the put of a type declaration into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the type declaration's references are stored as well.

putTypeDeclarationComponent :: Hash -> [Decl v a] -> Transaction ()

getTermComponentWithTypes :: Hash -> Transaction (Maybe [(Term v a, Type v a)])

getRootBranch :: m (Branch m)

Get the root branch.

putRootBranch :: Text -> Branch m -> m ()

Like $sel:putBranch:Codebase, but also adjusts the root branch pointer afterwards.

getBranchForHash :: CausalHash -> m (Maybe (Branch m))

putBranch :: Branch m -> m ()

Put a branch into the codebase, which includes its children, its patches, and the branch itself, if they don't already exist.

The terms and type declarations that a branch references must already exist in the codebase.

syncFromDirectory :: CodebasePath -> SyncMode -> Branch m -> m ()

Copy a branch and all of its dependencies from the given codebase into this one.

syncToDirectory :: CodebasePath -> SyncMode -> Branch m -> m ()

Copy a branch and all of its dependencies from this codebase into the given codebase.

viewRemoteBranch' :: forall r. ReadGitRemoteNamespace -> GitBranchBehavior -> ((Branch m, CodebasePath) -> m r) -> m (Either GitError r)

pushGitBranch :: forall e. WriteGitRepo -> PushGitBranchOpts -> (Branch m -> m (Either e (Branch m))) -> m (Either GitError (Either e (Branch m)))

Push the given branch to the given repo, and optionally set it as the root branch.

getWatch :: WatchKind -> Id -> Transaction (Maybe (Term v a))

getWatch k r returns watch result t that was previously put by putWatch k r t.

termsOfTypeImpl :: Reference -> Transaction (Set Id)

Get the set of user-defined terms-or-constructors that have the given type.

termsMentioningTypeImpl :: Reference -> Transaction (Set Id)

Get the set of user-defined terms-or-constructors mention the given type anywhere in their signature.

filterTermsByReferenceIdHavingTypeImpl :: TypeReference -> Set Id -> Transaction (Set Id)

Return the subset of the given set that has the given type.

filterTermsByReferentIdHavingTypeImpl :: TypeReference -> Set Id -> Transaction (Set Id)

Return the subset of the given set that has the given type.

termReferentsByPrefix :: ShortHash -> Transaction (Set Id)

Get the set of user-defined terms-or-constructors whose hash matches the given prefix.

withConnection :: forall x. (Connection -> m x) -> m x

Acquire a new connection to the same underlying database file this codebase object connects to.

withConnectionIO :: forall x. (Connection -> IO x) -> IO x

Acquire a new connection to the same underlying database file this codebase object connects to.

type CodebasePath = FilePath

A directory that contains a codebase.

data PushGitBranchOpts

Constructors

PushGitBranchOpts

Fields

behavior :: GitPushBehavior

syncMode :: SyncMode

data GitPushBehavior

Constructors

GitPushBehaviorGist
Don't set root, just sync entities.

GitPushBehaviorFf
After syncing entities, do a fast-forward check, then set the root.

GitPushBehaviorForce
After syncing entities, just set the root (force-pushy).

data GitError

Constructors

GitProtocolError GitProtocolError

GitCodebaseError (GitCodebaseError CausalHash)

GitSqliteCodebaseError GitSqliteCodebaseError

type SyncToDir m = CodebasePath -> SyncMode -> Branch m -> m ()

data LocalOrRemote

Whether a codebase is local or remote.

Constructors

Local

Remote

gitErrorFromOpenCodebaseError :: CodebasePath -> ReadGitRepo -> OpenCodebaseError -> GitSqliteCodebaseError