Unison.CommandLine

Pretty Printing

allow :: FilePath -> Bool

backtick :: IsString s => Pretty s -> Pretty s

aside :: (ListLike s Char, IsString s) => Pretty s -> Pretty s -> Pretty s

bigproblem :: (ListLike s Char, IsString s) => Pretty s -> Pretty s

note :: (ListLike s Char, IsString s) => Pretty s -> Pretty s

nothingTodo :: (ListLike s Char, IsString s) => Pretty s -> Pretty s

plural :: Foldable f => f a -> b -> b -> b

plural' :: Integral a => a -> b -> b -> b

problem :: (ListLike s Char, IsString s) => Pretty s -> Pretty s

tip :: (ListLike s Char, IsString s) => Pretty s -> Pretty s

warn :: (ListLike s Char, IsString s) => Pretty s -> Pretty s

warnNote :: String -> String

Completers

completion :: String -> Completion

completion' :: String -> Completion

exactComplete :: String -> [String] -> [Completion]

Constructs a list of Completions from a query and completion options by filtering them for prefix matches. A completion will be selected if it's an exact match for a provided option.

fuzzyComplete :: String -> [String] -> [Completion]

fuzzyCompleteHashQualified :: Names -> String -> [Completion]

prefixIncomplete :: String -> [String] -> [Completion]

prettyCompletion :: Bool -> (String, Pretty ColorText) -> Completion

fixupCompletion :: String -> [Completion] -> [Completion]

completeWithinQueryNamespace :: String -> [String] -> [Completion]

Completes a list of options, limiting options to the same namespace as the query, or the namespace's children if the query is itself a namespace.

E.g. query: "base" would match: ["base", "base.List", "base2"] wouldn't match: ["base.List.map", "contrib", "base2.List"]

Other

parseInput

Arguments

:: Branch0 m
Root branch, used to expand globs

-> Absolute
Current path from root, used to expand globs

-> [String]
Numbered arguments

-> Map String InputPattern
Input Pattern Map

-> [String]
command:arguments

-> Either (Pretty ColorText) Input

prompt :: String

watchBranchUpdates :: IO (Branch IO) -> TQueue Event -> Codebase IO v a -> IO (IO ())

watchConfig :: FilePath -> IO (Config, IO ())

watchFileSystem :: TQueue Event -> FilePath -> IO (IO ())

Exported for testing

beforeHash :: forall m h e. Monad m => Word -> RawHash h -> Causal m h e -> m Bool