Unison.CommandLine.Globbing

Description

Provides Globbing for selecting types, terms and namespaces using wildcards.

Documentation

expandGlobs

Arguments

:: forall m. Set TargetType

-> Branch0 m
Root branch

-> Absolute
UCM's current path

-> String
The glob string, e.g. .base.List.?.doc

-> [String]
Fully expanded, absolute paths. E.g. [".base.List.map"]

Expand a single glob pattern into all matching targets of the specified types.

data TargetType

Possible targets which a glob may select.

Constructors

Type

Term

Namespace