Unison.CommandLine.InputPatterns

Documentation

showPatternHelp :: InputPattern -> Pretty ColorText

patternName :: InputPattern -> Pretty ColorText

makeExample :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

makeExampleNoBackticks :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

makeExample' :: InputPattern -> Pretty ColorText

makeExampleEOS :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

helpFor :: InputPattern -> Either (Pretty ColorText) Input

mergeBuiltins :: InputPattern

mergeIOBuiltins :: InputPattern

updateBuiltins :: InputPattern

todo :: InputPattern

load :: InputPattern

add :: InputPattern

previewAdd :: InputPattern

update :: InputPattern

previewUpdate :: InputPattern

patch :: InputPattern

view :: InputPattern

display :: InputPattern

displayTo :: InputPattern

docs :: InputPattern

ui :: InputPattern

undo :: InputPattern

viewByPrefix :: InputPattern

find :: InputPattern

findShallow :: InputPattern

findVerbose :: InputPattern

findPatch :: InputPattern

renameTerm :: InputPattern

renameType :: InputPattern

delete :: InputPattern

deleteTerm :: InputPattern

deleteType :: InputPattern

deleteTermReplacementCommand :: String

deleteTypeReplacementCommand :: String

deleteReplacement :: Bool -> InputPattern

deleteTermReplacement :: InputPattern

deleteTypeReplacement :: InputPattern

parseHashQualifiedName :: String -> Either (Pretty ColorText) (HashQualified Name)

aliasTerm :: InputPattern

aliasType :: InputPattern

aliasMany :: InputPattern

up :: InputPattern

cd :: InputPattern

back :: InputPattern

deleteBranch :: InputPattern

deletePatch :: InputPattern

movePatch :: String -> String -> Either (Pretty ColorText) Input

copyPatch' :: String -> String -> Either (Pretty ColorText) Input

copyPatch :: InputPattern

renamePatch :: InputPattern

renameBranch :: InputPattern

history :: InputPattern

forkLocal :: InputPattern

resetRoot :: InputPattern

pullSilent :: InputPattern

pull :: InputPattern

pullImpl :: String -> Verbosity -> InputPattern

pullExhaustive :: InputPattern

push :: InputPattern

pushExhaustive :: InputPattern

createPullRequest :: InputPattern

loadPullRequest :: InputPattern

parseUri :: String -> String -> Either (Pretty ColorText) ReadRemoteNamespace

parseWriteRepo :: String -> String -> Either (Pretty ColorText) WriteRepo

parsePushPath :: String -> String -> Either (Pretty ColorText) WriteRemotePath

squashMerge :: InputPattern

mergeLocal :: InputPattern

diffNamespace :: InputPattern

previewMergeLocal :: InputPattern

replaceEdit :: (HashQualified Name -> HashQualified Name -> Maybe PatchPath -> Input) -> InputPattern

replace :: InputPattern

viewReflog :: InputPattern

edit :: InputPattern

topicNameArg :: ArgumentType

helpTopics :: InputPattern

helpTopicsMap :: Map String (Pretty ColorText)

help :: InputPattern

quit :: InputPattern

viewPatch :: InputPattern

link :: InputPattern

links :: InputPattern

unlink :: InputPattern

names :: InputPattern

dependents :: InputPattern

dependencies :: InputPattern

debugNumberedArgs :: InputPattern

debugFileHashes :: InputPattern

debugDumpNamespace :: InputPattern

debugDumpNamespaceSimple :: InputPattern

debugClearWatchCache :: InputPattern

test :: InputPattern

execute :: InputPattern

ioTest :: InputPattern

makeStandalone :: InputPattern

createAuthor :: InputPattern

validInputs :: [InputPattern]

commandNames :: [String]

commandNameArg :: ArgumentType

exactDefinitionOrPathArg :: ArgumentType

fuzzyDefinitionQueryArg :: ArgumentType

definitionQueryArg :: ArgumentType

exactDefinitionTypeQueryArg :: ArgumentType

exactDefinitionTermQueryArg :: ArgumentType

typeCompletor :: Applicative m => (String -> [String] -> [Completion]) -> String -> Codebase m v a -> Branch m -> Absolute -> m [Completion]

termCompletor :: Applicative m => (String -> [String] -> [Completion]) -> String -> Codebase m v a -> Branch m -> Absolute -> m [Completion]

patchArg :: ArgumentType

allCompletors :: Monad m => [String -> Codebase m v a -> Branch m -> Absolute -> m [Completion]] -> String -> Codebase m v a -> Branch m -> Absolute -> m [Completion]

bothCompletors :: Monad m => (String -> t2 -> t3 -> t4 -> m [Completion]) -> (String -> t2 -> t3 -> t4 -> m [Completion]) -> String -> t2 -> t3 -> t4 -> m [Completion]

pathCompletor

Arguments

:: Applicative f

=> (String -> [String] -> [Completion])
Turns a query and list of possible completions into a Completion.

-> (Branch0 m -> Set Text)
Construct completions given ucm's current branch context, or the root namespace if the query is absolute.

-> String
The portion of this arg that the user has already typed.

-> codebase

-> Branch m

-> Absolute

-> f [Completion]

namespaceArg :: ArgumentType

allSubNamespaces :: Branch0 m -> [Text]

Recursively collects all names of namespaces which are children of the branch.

newNameArg :: ArgumentType

noCompletions :: ArgumentType

gitUrlArg :: ArgumentType

collectNothings :: (a -> Maybe b) -> [a] -> [a]

patternFromInput :: Input -> InputPattern

inputStringFromInput :: IsString s => Input -> Pretty s