Unison.FileParsers

Documentation

data ShouldUseTndr m

Should we use type-directed name resolution?

Constructors

ShouldUseTndr'No

ShouldUseTndr'Yes (ParsingEnv m)

computeTypecheckingEnvironment :: forall m v. (Var v, Monad m) => ShouldUseTndr m -> [Type v] -> (DefnsF Set TermReference TypeReference -> m (TypeLookup v Ann)) -> UnisonFile v -> m (Env v Ann)

Compute a typechecking environment, given:

Whether or not to use type-directed name resolution during type checking.

The abilities that are considered to already have ambient handlers.

A function to compute a TypeLookup for the given set of type- or term-references.

The parsing environment that was used to parse the parsed Unison file.

The parsed Unison file for which the typechecking environment is applicable.

synthesizeFile :: forall m v. (Monad m, Var v) => Env v Ann -> UnisonFile v -> ResultT (Seq (Note v Ann)) m (TypecheckedUnisonFile v Ann)