Unison.Hashing.V2.ABT

Documentation

data Term (f :: Type -> Type) v a

At each level in the tree, we store the set of free variables and a value of type a. Variables are of type v.

hash :: forall f v a h. (Functor f, Hashable1 f, Eq v, Show v, Ord h, Accumulate h) => Term f v a -> h

We ignore annotations in the Term, as these should never affect the meaning of the term.

hashComponents :: (Functor f, Hashable1 f, Foldable f, Eq v, Show v, Var v, Ord h, Accumulate h) => (h -> Word64 -> Term f v ()) -> Map v (Term f v a) -> [(h, [(v, Term f v a)])]