Unison.Hashing.V2.DataDeclaration

Documentation

data DataDeclaration v a

Constructors

DataDeclaration

Fields

modifier :: Modifier

annotation :: a

bound :: [v]

constructors' :: [(a, v, Type v a)]

newtype EffectDeclaration v a

Constructors

EffectDeclaration

Fields

toDataDecl :: DataDeclaration v a

type Decl v a = Either (EffectDeclaration v a) (DataDeclaration v a)

data Modifier

Constructors

Structural

Unique Text

asDataDecl :: Decl v a -> DataDeclaration v a

constructorType :: Decl v a -> ConstructorType

constructorTypes :: DataDeclaration v a -> [Type v a]

declDependencies :: Ord v => Decl v a -> Set Reference

dependencies :: Ord v => DataDeclaration v a -> Set Reference

bindReferences :: Var v => Set v -> Map Name Reference -> DataDeclaration v a -> ResolutionResult v a (DataDeclaration v a)

hashDecls :: (Eq v, Var v, Show v) => Map v (DataDeclaration v a) -> ResolutionResult v a [(v, Id, DataDeclaration v a)]

compute the hashes of these user defined types and update any free vars corresponding to these decls with the resulting hashes

data List a = Nil | Cons a (List a) becomes something like (List, #xyz, [forall a. #xyz a, forall a. a -> (#xyz a) -> (#xyz a)])

NOTE: technical limitation, this implementation gives diff results if ctors have the same FQN as one of the types. TODO: assert this and bomb if not satisfied, or else do local mangling and unmangling to ensure this doesn't affect the hash.