Unison.Hashing.V2.Hashable

Documentation

class Hashable t where

This typeclass provides a mechanism for obtaining a content-based hash for Unison types & terms. Be wary that Unison requires that these hashes be deterministic, any change to a Hashable instance requires a full codebase migration and should not be taken lightly.

Methods

hash :: t -> Hash

hashFor :: Hashable t => t -> HashFor t

newtype HashFor t

A hash tagged with the type it's a hash of, useful for maintaining type safety guarantees.

Constructors

HashFor

Fields

genericHash :: Hash