Unison.Hashing.V2.Term

Documentation

type Term v a = Term2 v a a v a

Like `Term v`, but with an annotation of type a at every level in the tree

data F typeVar typeAnn patternAnn a

Base functor for terms in the Unison language We need typeVar because the term and type variables may differ.

Constructors

Int Int64

Nat Word64

Float Double

Boolean Bool

Text Text

Char Char

Blank (Blank typeAnn)

Ref Reference

Constructor Reference ConstructorId

Request Reference ConstructorId

Handle a a

App a a

Ann a (Type typeVar typeAnn)

List (Seq a)

If a a a

And a a

Or a a

Lam a

LetRec IsTop [a] a

Let IsTop a a

Match a [MatchCase patternAnn a]

TermLink Referent

TypeLink Reference

data MatchCase loc a

Constructors

MatchCase (Pattern loc) (Maybe a) a

hashComponents :: Var v => Map v (Term v a) -> Map v (Id, Term v a)

hashClosedTerm :: Var v => Term v a -> Id