Unison.Hashing.V2.Tokenizable

Documentation

class Tokenizable t where

Tokenizable converts a value into a set of hashing tokens which will later be accumulated into a Hash. Be very careful when adding or altering instances of this typeclass, changing the hash of a value is a major breaking change and requires a complete codebase migration.

If you simply want to provide a convenience instance for a type which wraps some Hashable type, write an instance of Hashable which calls through to the inner instance instead.

E.g. If I want to be able to hash a TaggedBranch using its Branch0 hashable instance:

@@ data TaggedBranch = TaggedBranch String Branch

instance Hashable TaggedBranch where hash (TaggedBranch _ b) = hash b @@

Methods

tokens :: Accumulate h => t -> [Token h]

class Accumulate h where

Methods

accumulate :: [Token h] -> h

fromBytes :: ByteString -> h

toBytes :: h -> ByteString

class Hashable1 f where

Methods

hash1 :: (Ord h, Accumulate h) => ([a] -> ([h], a -> h)) -> (a -> h) -> f a -> h

Produce a hash for an `f a`, given a hashing function for a. If there is a notion of order-independence in some aspect of a subterm of f, then the first argument (`hashUnordered :: [a] -> ([h], a -> h)`) should be used to impose an order, and then apply that order in further hashing. Otherwise the second argument (`hash :: a -> h`) should be used.

Example 1: A simple functor with no unordered components. Hashable1 instance just uses hash:

data T a = One a | Two a a deriving Functor

instance Hashable1 T where hash1 _ hash t = case t of One a -> accumulate [Tag 0, Hashed (hash a)] Two a a2 -> accumulate [Tag 1, Hashed (hash a), Hashed (hash a2)]

Example 2: A functor with unordered components. For hashing, we need to pick a canonical ordering of the unordered components, so we use hashUnordered:

data U a = U { unordered :: [a], uno :: a, dos :: a } deriving Functor

instance Hashable1 U where hash1 hashUnordered _ (U unordered uno dos) = let (hs, hash) = hashUnordered unordered in accumulate $ map Hashed hs ++ [Hashed (hash uno), Hashed (hash dos)]

data Token h

Constructors

Tag !Word8

Bytes !ByteString

Int !Int64

Text !Text

Double !Double

Hashed !h

Nat !Word64

hashTokenizable :: (Tokenizable t, Accumulate h) => t -> h

Tokenize then accumulate a type into a Hash.

accumulateToken :: (Accumulate h, Tokenizable t) => t -> Token h