Unison.KindInference.Constraint.Pretty

Description

Documentation

prettyUVarKind :: Var v => PrettyPrintEnv -> ConstraintMap v loc -> UVar v loc -> Pretty ColorText

Pretty print the kind constraint on the given UVar.

Precondition: The ConstraintMap is acyclic.

prettySolvedConstraint :: Var v => PrettyPrintEnv -> ConstraintMap v loc -> Constraint (UVar v loc) v loc -> Pretty ColorText

Pretty print a Constraint

Precondition: The ConstraintMap is acyclic.

prettyCyclicUVarKind

Arguments

:: Var v

=> PrettyPrintEnv

-> ConstraintMap v loc

-> UVar v loc

-> (Pretty ColorText -> Pretty ColorText)
A function to style the cyclic UVar's variable name

-> (Pretty ColorText, Pretty ColorText)
(the pretty UVar variable, the generating equation)

A pretty printer for cyclic kind constraints on a UVar. Expresses the infinite kind by a generating equation.

Precondition: The UVar has a cyclic constraint.