Unison.KindInference.Constraint.Provenance

Documentation

data Provenance v loc

A tag for a source span and context indicating where the generated constraint came from. This is helpful to propagate the constraint context through the solver for user-facing error messages.

Constructors

Provenance !(ConstraintContext v loc) !loc

loc :: Lens' (Provenance v loc) loc