Unison.KindInference.Error

Documentation

data KindError v loc

Errors that may arise during kind inference

Constructors

CycleDetected loc (UVar v loc) (ConstraintMap v loc)
A variable is constrained to have an infinite kind

UnexpectedArgument
Something of kind * or Effect is applied to an argument

Fields

loc

src span of abs

(UVar v loc)

abs var

(UVar v loc)

arg var

(ConstraintMap v loc)

context | An arrow kind is applied to a type, but its kind doesn't match the expected argument kind

ArgumentMismatch

Fields

(UVar v loc)

abs var

(UVar v loc)

expected var

(UVar v loc)

given var

(ConstraintMap v loc)

context | Same as ArgumentMismatch, but for applications to the builtin Arrow type.

ArgumentMismatchArrow

Fields

(loc, Type v loc, Type v loc)

(The applied arrow range, lhs, rhs)

(ConstraintConflict v loc)

(ConstraintMap v loc)

EffectListMismatch (ConstraintConflict v loc) (ConstraintMap v loc)
Something appeared in an effect list that isn't of kind Effect

ConstraintConflict
Generic constraint conflict

Fields

(GeneratedConstraint v loc)

Failed to add this constraint

(ConstraintConflict v loc)

Due to this conflict

(ConstraintMap v loc)

in this context

lspLoc :: Semigroup loc => KindError v loc -> loc

data ConstraintConflict v loc

Two incompatible constraints on a UVar.

Constructors

ConstraintConflict'

Fields

conflictedVar :: UVar v loc

impliedConstraint :: Constraint (UVar v loc) v loc

conflictedConstraint :: Constraint (UVar v loc) v loc

improveError :: Var v => KindError v loc -> Solve v loc (KindError v loc)

Transform generic constraint conflicts into more specific error by examining its ConstraintContext.