Unison.PatternMatchCoverage.Class

Documentation

class (Ord loc, Var vt, Var v, MonadFix m) => Pmc vt v loc m | m -> vt v loc where

A typeclass for the queries required to perform pattern match coverage checking.

Methods

getConstructors :: Type vt loc -> m (EnumeratedConstructors vt v loc)

Get the constructors of a type

getConstructorVarTypes :: Type vt loc -> ConstructorReference -> m [Type vt loc]

Get the types of the arguments of a specific constructor

fresh :: m v

Get a fresh variable

getPrettyPrintEnv :: m PrettyPrintEnv

data EnumeratedConstructors vt v loc

Constructors

ConstructorType [(v, ConstructorReference, Type vt loc)]

AbilityType (Type vt loc) (Map ConstructorReference (v, Type vt loc))

SequenceType [(ListPat, [Type vt loc])]

BooleanType

OtherType

traverseConstructorTypes :: Applicative f => (v -> ConstructorReference -> Type vt loc -> f (Type vt loc)) -> EnumeratedConstructors vt v loc -> f (EnumeratedConstructors vt v loc)