Unison.PatternMatchCoverage.PmGrd

Documentation

data PmGrd vt v loc

Constructors

PmCon
PmCon x Con xs ys corresponds to the constraint Con ys <- x

Fields

v

Variable

ConstructorReference

Constructor

[(v, Type vt loc)]

Constructor argument values and types

PmEffect

Fields

v

Variable

ConstructorReference

Constructor

[(v, Type vt loc)]

Constructor argument values and types

PmEffectPure v (v, Type vt loc)

PmLit v PmLit

PmListHead

Fields

v

list root

Int

cons position (0 is head)

v

element variable

(Type vt loc)

element type

PmListTail

Fields

v

list root

Int

snoc position (0 is last)

v

element variable

(Type vt loc)

element type

PmListInterval v Int Int
The size of the list must fall within this inclusive range

PmBang v
If a guard performs an effect

PmLet v (Term' vt v loc) (Type vt loc)
PmLet x expr corresponds to a let x = expr guard. This actually binds x.

prettyPmGrd :: (Var vt, Var v) => PrettyPrintEnv -> PmGrd vt v loc -> Pretty ColorText