Unison.PrettyPrintEnv

Documentation

data PrettyPrintEnv

Constructors

PrettyPrintEnv

Fields

termNames :: Referent -> [(HashQualified Name, HashQualified Name)]

typeNames :: Reference -> [(HashQualified Name, HashQualified Name)]

patterns :: PrettyPrintEnv -> ConstructorReference -> Maybe (HashQualified Name)

patternName :: PrettyPrintEnv -> ConstructorReference -> HashQualified Name

terms :: PrettyPrintEnv -> Referent -> Maybe (HashQualified Name)

types :: PrettyPrintEnv -> Reference -> Maybe (HashQualified Name)

allTermNames :: PrettyPrintEnv -> Referent -> [HashQualified Name]

allTypeNames :: PrettyPrintEnv -> Reference -> [HashQualified Name]

termName :: PrettyPrintEnv -> Referent -> HashQualified Name

typeName :: PrettyPrintEnv -> Reference -> HashQualified Name

termNameOrHashOnly :: PrettyPrintEnv -> Referent -> HashQualified Name

typeNameOrHashOnly :: PrettyPrintEnv -> Reference -> HashQualified Name

biasTo :: [Name] -> PrettyPrintEnv -> PrettyPrintEnv

Prefer names which share a common prefix with any provided target.

Results are sorted according to the longest common prefix found against ANY target.

labeledRefName :: PrettyPrintEnv -> LabeledDependency -> HashQualified Name

Get a name for a LabeledDependency from the PPE.

Exported only for cases where the codebase's configured hash length is unavailable.

todoHashLength :: Int

addFallback :: PrettyPrintEnv -> PrettyPrintEnv -> PrettyPrintEnv

Attempts to find a name in primary ppe, falls back to backup ppe only if no names are found. Typically one can use this to shadow global or absolute names with names that are within the current path.

union :: PrettyPrintEnv -> PrettyPrintEnv -> PrettyPrintEnv

Finds names from both PPEs, if left unbiased the name from the left ppe is preferred.

This is distinct from addFallback with respect to biasing; A bias applied to a union might select a name in the right half of the union. Whereas, a bias applied to the result of addFallback will bias within the available names inside the left PPE and will only search in the fallback if there aren't ANY names in the primary ppe.

If you don't know the difference, it's likely you want addFallback where you add global names as a fallback for local names.

empty :: PrettyPrintEnv