Unison.PrettyPrintEnv.MonadPretty

Documentation

type MonadPretty v (m :: Type -> Type) = (Var v, MonadReader (Env v) m)

data Env v

Constructors

Env

Fields

boundTerms :: !(Set v)

boundTypes :: !(Set v)

freeTerms :: !(Set v)

ppe :: !PrettyPrintEnv

forceRawStrings :: !Bool

When True, always use raw strings (triple-quoted) for multiline text, even when nested inside other expressions. This is useful for diff output where we want actual newlines for better line-by-line diffing.

runPretty :: Var v => PrettyPrintEnv -> Reader (Env v) a -> a

runPrettyForDiff :: Var v => PrettyPrintEnv -> Reader (Env v) a -> a

Like runPretty, but enables raw string rendering for multiline text. This is useful for diff output where we want actual newlines for better diffing.

addTypeVars :: MonadPretty v m => [v] -> m a -> m a

Add type variables to the set of variables that need to be avoided

willCaptureType :: MonadPretty v m => [v] -> m Bool

Check if a list of type variables contains any variables that need to be avoided

withBoundTerm :: MonadPretty v m => v -> m a -> m a

withBoundTerms :: MonadPretty v m => [v] -> m a -> m a