Unison.Project.Util

Documentation

projectPath :: ProjectId -> Absolute

Get the path that a project is stored at. Users aren't supposed to go here.

>>> projectPath "ABCD"
.__projects._ABCD

projectBranchesPath :: ProjectId -> Absolute

Get the path that a project's branches are stored at. Users aren't supposed to go here.

>>> projectBranchesPath "ABCD"
.__projects._ABCD.branches

projectBranchPath :: ProjectAndBranch ProjectId ProjectBranchId -> Absolute

Get the path that a branch is stored at. Users aren't supposed to go here.

>>> projectBranchPath ProjectAndBranch { project = "ABCD", branch = "DEFG" }
.__projects._ABCD.branches._DEFG

projectBranchSegment :: ProjectBranchId -> NameSegment

Get the name segment that a branch is stored at.

>>> projectBranchSegment "DEFG"
"_DEFG"

projectPathPrism :: Prism' Absolute ProjectId

The prism between paths like

.__projects._XX_XX

and the project id

XX-XX

projectBranchPathPrism :: Prism' Absolute (ProjectAndBranch ProjectId ProjectBranchId, Path)

The prism between paths like

.__projects._XX_XX.branches._YY_YY.foo.bar

and the (project id, branch id, path) triple

(XX-XX, YY-YY, foo.bar)

projectContextFromPath :: Absolute -> ProjectContext

pattern UUIDNameSegment :: UUID -> NameSegment

data ProjectContext

The project information about the current path. NOTE: if the user has cd'd into the project storage area but NOT into a branch, (where they shouldn't ever be), this will result in a LooseCodePath.

Constructors

LooseCodePath Absolute

ProjectBranchPath ProjectId ProjectBranchId Path

pattern ProjectsNameSegment :: NameSegment

pattern BranchesNameSegment :: NameSegment