Unison.Server.Backend

Documentation

type SyntaxText = SyntaxText' Reference

data ShallowListEntry v a

Constructors

ShallowTermEntry (TermEntry v a)

ShallowTypeEntry TypeEntry

ShallowBranchEntry NameSegment ShortBranchHash Int

ShallowPatchEntry NameSegment

listEntryName :: ShallowListEntry v a -> Text

data BackendError

Constructors

NoSuchNamespace Absolute

BadRootBranch GetRootBranchError

CouldntExpandBranchHash ShortBranchHash

AmbiguousBranchHash ShortBranchHash (Set ShortBranchHash)

NoBranchForHash Hash

CouldntLoadBranch Hash

MissingSignatureForTerm Reference

type Backend m a = ExceptT BackendError m a

basicNames' :: Branch m -> Path -> (Names, Names)

basicSuffixifiedNames :: Int -> Branch m -> Path -> PrettyPrintEnv

basicPrettyPrintNames :: Branch m -> Path -> Names

basicParseNames :: Branch m -> Path -> Names

loadReferentType :: (Applicative m, Var v) => Codebase m v Ann -> Referent -> m (Maybe (Type v Ann))

getRootBranch :: Functor m => Codebase m v Ann -> Backend m (Branch m)

data TermEntry v a

Constructors

TermEntry

Fields

termEntryReferent :: Referent

termEntryName :: HQSegment

termEntryType :: Maybe (Type v a)

termEntryTag :: Maybe TermTag

data TypeEntry

Constructors

TypeEntry

Fields

typeEntryReference :: Reference

typeEntryName :: HQSegment

typeEntryTag :: TypeTag

data FoundRef

Constructors

FoundTermRef Referent

FoundTypeRef Reference

fuzzyFind :: Monad m => Path -> Branch m -> String -> [(Alignment, UnisonName, [FoundRef])]

findShallow :: (Monad m, Var v) => Codebase m v Ann -> Absolute -> Backend m [ShallowListEntry v Ann]

findShallowReadmeInBranchAndRender :: Var v => Width -> Runtime v -> Codebase IO v Ann -> Branch IO -> Backend IO (Maybe Doc)

termListEntry :: Monad m => Var v => Codebase m v Ann -> Branch0 m -> Referent -> HQSegment -> Backend m (TermEntry v Ann)

typeListEntry :: Monad m => Var v => Codebase m v Ann -> Reference -> HQSegment -> Backend m TypeEntry

typeDeclHeader :: forall v m. Monad m => Var v => Codebase m v Ann -> PrettyPrintEnv -> Reference -> Backend m (DisplayObject SyntaxText SyntaxText)

formatTypeName :: PrettyPrintEnv -> Reference -> SyntaxText

formatTypeName' :: PrettyPrintEnv -> Reference -> SyntaxText

termEntryToNamedTerm :: Var v => PrettyPrintEnv -> Maybe Width -> TermEntry v a -> NamedTerm

typeEntryToNamedType :: TypeEntry -> NamedType

findShallowInBranch :: (Monad m, Var v) => Codebase m v Ann -> Branch m -> Backend m [ShallowListEntry v Ann]

termReferencesByShortHash :: Monad m => Codebase m v a -> ShortHash -> m (Set Reference)

typeReferencesByShortHash :: Monad m => Codebase m v a -> ShortHash -> m (Set Reference)

termReferentsByShortHash :: Monad m => Codebase m v a -> ShortHash -> m (Set Referent)

getCurrentPrettyNames :: Path -> Branch m -> NamesWithHistory

getCurrentParseNames :: Path -> Branch m -> NamesWithHistory

fixupNamesRelative :: Absolute -> Names -> Names

data Search r

A Search r is a small bag of functions that is used to power a search for rs.

Construct a Search with makeTypeSearch or makeTermSearch, and eliminate it with applySearch.

Constructors

Search

Fields

lookupNames :: r -> Set (HashQualified Name)

lookupRelativeHQRefs' :: HashQualified Name -> Set r

makeResult :: HashQualified Name -> r -> Set (HashQualified Name) -> SearchResult

matchesNamedRef :: Name -> r -> HashQualified Name -> Bool

makeTypeSearch :: Int -> NamesWithHistory -> Search Reference

Make a type search, given a short hash length and names to search in.

makeTermSearch :: Int -> NamesWithHistory -> Search Referent

Make a term search, given a short hash length and names to search in.

applySearch :: Show r => Search r -> HashQualified Name -> [SearchResult]

Interpret a Search as a function from name to search results.

hqNameQuery :: Monad m => Maybe Path -> Branch m -> Codebase m v Ann -> [HashQualified Name] -> m QueryResult

data DefinitionResults v

Constructors

DefinitionResults

Fields

termResults :: Map Reference (DisplayObject (Type v Ann) (Term v Ann))

typeResults :: Map Reference (DisplayObject () (Decl v Ann))

noResults :: [HashQualified Name]

collateReferences :: Foldable f => Foldable g => f Reference -> g Referent -> (Set Reference, Set Reference)

expandShortBranchHash :: Monad m => Codebase m v a -> ShortBranchHash -> Backend m Hash

formatType' :: Var v => PrettyPrintEnv -> Width -> Type v a -> SyntaxText

formatType :: Var v => PrettyPrintEnv -> Width -> Type v a -> SyntaxText

formatSuffixedType :: Var v => PrettyPrintEnvDecl -> Width -> Type v Ann -> SyntaxText

mungeSyntaxText :: Functor g => g (Element Reference) -> g Element

prettyDefinitionsBySuffixes :: forall v. Var v => Maybe Path -> Maybe Hash -> Maybe Width -> Suffixify -> Runtime v -> Codebase IO v Ann -> [HashQualified Name] -> Backend IO DefinitionDisplayResults

renderDoc :: forall v. Var v => PrettyPrintEnvDecl -> Width -> Runtime v -> Codebase IO v Ann -> Reference -> Backend IO [(HashQualifiedName, UnisonHash, Doc)]

bestNameForTerm :: forall v. Var v => PrettyPrintEnv -> Width -> Referent -> Text

bestNameForType :: forall v. Var v => PrettyPrintEnv -> Width -> Reference -> Text

resolveBranchHash :: Monad m => Maybe Hash -> Codebase m v Ann -> Backend m (Branch m)

resolveRootBranchHash :: Monad m => Maybe ShortBranchHash -> Codebase m v Ann -> Backend m (Branch m)

definitionsBySuffixes :: forall m v. MonadIO m => Var v => Maybe Path -> Branch m -> Codebase m v Ann -> [HashQualified Name] -> Backend m (DefinitionResults v)

termsToSyntax :: Var v => Ord a => Suffixify -> Width -> PrettyPrintEnvDecl -> Map Reference (DisplayObject (Type v a) (Term v a)) -> Map Reference (DisplayObject SyntaxText SyntaxText)

typesToSyntax :: Var v => Ord a => Suffixify -> Width -> PrettyPrintEnvDecl -> Map Reference (DisplayObject () (Decl v a)) -> Map Reference (DisplayObject SyntaxText SyntaxText)

loadSearchResults :: (Var v, Applicative m) => Codebase m v Ann -> [SearchResult] -> m [SearchResult' v Ann]

loadTypeDisplayObject :: Applicative m => Codebase m v Ann -> Reference -> m (DisplayObject () (Decl v Ann))