Unison.Share.Types

Description

Types related to Share and Codeservers.

Documentation

data CodeserverURI

This type is expanded out into all of its fields because we require certain pieces which are optional in a URI, and also to make it more typesafe to eventually convert into a BaseURL for servant clients.

Constructors

CodeserverURI

Fields

codeserverScheme :: Scheme

codeserverUserInfo :: String

codeserverRegName :: String

codeserverPort :: Maybe Int

codeserverPath :: [String]

newtype CodeserverId

This is distinct from the codeserver URI in that we store credentials by a normalized ID, since it's much easier to look up that way than from an arbitrary path. We may wish to use explicitly named configurations in the future. This currently uses a stringified uriAuthority.

Constructors

CodeserverId

Fields

codeserverId :: Text

data Scheme

Constructors

Http

Https

codeserverFromURI :: URI -> Maybe CodeserverURI

>>> import Data.Maybe (fromJust)
>>> codeserverFromURI . fromJust $ parseURI "http://localhost:8080"
Just http://localhost:8080
>>> codeserverFromURI . fromJust $ parseURI "http://localhost:80"
Just http://localhost:80
>>> codeserverFromURI . fromJust $ parseURI "https://share.unison-lang.org/api"
Just https://share.unison-lang.org/api
>>> codeserverFromURI . fromJust $ parseURI "http://share.unison-lang.org/api"
Just http://share.unison-lang.org/api

codeserverIdFromURI :: URI -> Either Text CodeserverId

Gets the part of the CodeserverURI that we use for identifying that codeserver in credentials files.

>>> import Data.Maybe (fromJust)
>>> import Network.URI (parseURI)
>>> codeserverIdFromURI (fromJust $ parseURI "http://localhost:5424/api")
>>> codeserverIdFromURI (fromJust $ parseURI "https://share.unison-lang.org/api")
Right "localhost:5424"
Right "share.unison-lang.org"

codeserverToURI :: CodeserverURI -> URI

codeserverIdFromCodeserverURI :: CodeserverURI -> CodeserverId

Gets the CodeserverId for a given CodeserverURI

codeserverBaseURL :: CodeserverURI -> BaseUrl

Builds a servant-compatible BaseUrl for a given CodeserverURI.