Unison.Syntax.NamePrinter

Documentation

type SyntaxText = SyntaxText' Reference

prettyName :: IsString s => Name -> Pretty s

prettyHashQualified :: HashQualified Name -> Pretty SyntaxText

prettyHashQualified' :: HashQualified Name -> Pretty SyntaxText

prettyHashQualified0 :: IsString s => HashQualified Name -> Pretty s

prettyNamedReference :: Int -> Name -> Reference -> Pretty SyntaxText

Pretty-print a reference as a name and the given number of characters of its hash.

prettyNamedReferent :: Int -> Name -> Referent -> Pretty SyntaxText

Pretty-print a referent as a name and the given number of characters of its hash.

prettyReference :: Int -> Reference -> Pretty SyntaxText

Pretty-print a reference as the given number of characters of its hash.

prettyReferent :: Int -> Referent -> Pretty SyntaxText

Pretty-print a referent as the given number of characters of its hash.

prettyLabeledDependency :: Int -> LabeledDependency -> Pretty SyntaxText

prettyShortHash :: IsString s => ShortHash -> Pretty s

styleHashQualified :: IsString s => (Pretty s -> Pretty s) -> HashQualified Name -> Pretty s

styleHashQualified' :: IsString s => (Pretty s -> Pretty s) -> (Pretty s -> Pretty s) -> HashQualified Name -> Pretty s

styleHashQualified'' :: (Pretty SyntaxText -> Pretty SyntaxText) -> HashQualified Name -> Pretty SyntaxText

fmt :: Element r -> Pretty (SyntaxText' r) -> Pretty (SyntaxText' r)