Unison.Syntax.Precedence

Documentation

infixRules :: Map Text Precedence

isTopLevelPrecedence :: Precedence -> Bool

Indicates this is the RHS of a top-level definition.

increment :: Precedence -> Precedence

data Precedence

Constructors

Basement
The lowest precedence, used for top-level bindings

Bottom
Used for terms that never need parentheses

Annotation
Type annotations

Statement
A statement in a block

Control
Control flow constructs like `if`, match, `case`

InfixOp InfixPrecedence
Infix operators

Application
Function application

Prefix
Prefix operators like `'`, !

Top
The highest precedence, used for let bindings and blocks

data InfixPrecedence

Constructors

Lowest

Level Int

Highest

infixLevels :: [[Text]]

operatorPrecedence :: Text -> Maybe Precedence

Returns the precedence of an infix operator, if it has one.