Unison.Syntax.TermPrinter

Documentation

emptyAc :: AmbientContext

pretty :: Var v => PrettyPrintEnv -> Term v a -> Pretty ColorText

prettyBlock :: Var v => Bool -> PrettyPrintEnv -> Term v a -> Pretty ColorText

prettyBlock' :: Var v => Bool -> PrettyPrintEnv -> Term v a -> Pretty SyntaxText

pretty' :: Var v => Maybe Width -> PrettyPrintEnv -> Term v a -> ColorText

prettyBinding :: Var v => PrettyPrintEnv -> HashQualified Name -> Term2 v at ap v a -> Pretty SyntaxText

Render a binding, producing output of the form

foo : t -> u foo a = ...

The first line is only output if the term has a type annotation as the outermost constructor.

Binary functions with symbolic names are output infix, as follows:

(+) : t -> t -> t a + b = ...

prettyBinding' :: Var v => PrettyPrintEnv -> Width -> HashQualified Name -> Term v a -> ColorText

prettyBindingWithoutTypeSignature :: Var v => PrettyPrintEnv -> HashQualified Name -> Term2 v at ap v a -> Pretty SyntaxText

Like prettyBinding, but elides the type signature (if any).

prettyDoc2 :: forall v m. MonadPretty v m => AmbientContext -> Term3 v PrintAnnotation -> m (Maybe (Pretty SyntaxText))

pretty0 :: forall v m. MonadPretty v m => AmbientContext -> Term3 v PrintAnnotation -> m (Pretty SyntaxText)

runPretty :: Var v => PrettyPrintEnv -> Reader (PrettyPrintEnv, Set v) a -> a

prettyPattern :: forall v loc. Var v => PrettyPrintEnv -> AmbientContext -> Precedence -> [v] -> Pattern loc -> (Pretty SyntaxText, [v])