Unison.Typechecker.Context.Structure

Documentation

data Element v loc

Elements of an ordered algorithmic context. A context is a sequence of Elements (here represented as a finger tree), but not every sequence represents a valid context. A context is only valid if every type that occurs in the context is well scoped with regard to the previous variable bindings in the context.

We will refer to invalid trees as 'context segments' since they can be prefixed to become part of a valid context. Keeping track of the same context information for _every_ segment is important to avoiding some performance problems with a more naive implementation.

Constructors

Var (TypeVar v loc)
A variable declaration

Solved (Blank loc) v (Monotype v loc)
v is solved to some monotype

Ann v loc (Type v loc)
v has type a, maybe quantified loc contains the span of the name of the bound v

Marker v
used for scoping

variable :: Element v loc -> v

data Info v loc

Aggregate context (segment) information, built up from Elements.

boundExistentialVars is the set of existential variables that are known to be 'in scope' due to the context segment.

solvedExistentials is a mapping of existential variables to their solutions in the segment.

boundUniversalVars is the set of universal variables that are known to be 'in scope' due to the context segment.

termVarAnnotations is a mapping of term variables to their type.

allBoundVars contains all the variables 'in scope' due to the segment. Primarily used for freshening.

`freeUniversal/ExistentialVars` contains all the variables that occur _free_ in types in the segment. This means that they are _not_ bound in an Element previous to the one that the type occurs in.

shadowedVars contains variables that occur _multiple_ times in the segment, and thus the latter shadows the former.

recorded tracks certain annotated elements.

A context segment is a well formed context if `free*Vars` and shadowedVars are both empty.

Note: the fields here are intentionally lazy. In situations where there is heavy context manipulation, this seems to actually be faster, probably because the info is not actually used all of the time, and any fields that are strict are actually demanded by use of every other field in the info, regardless of whether the strict field ends up needed.

Constructors

Info

Fields

boundExistentialVars :: Set v

solvedExistentials :: Map v (Monotype v loc)

boundUniversalVars :: Set v

termVarAnnotations :: Map v (loc, Type v loc)

allBoundVars :: Set v

freeUniversalVars :: Set v

freeExistentialVars :: Set v

shadowedVars :: Set v

recorded :: Map v (Recorded loc, Type v loc)

unsolvedExistentials :: Ord v => Info v loc -> Set v

Gets the _unsolved_ existential variables of a context info.

newtype Context v loc

Constructors

Context (CtxSegment v loc)

context0 :: Ord v => Context v loc

The empty context

type CtxSegment v loc = FingerTree (Info v loc) (Element v loc)

(|>) :: Measured v a => FingerTree v a -> a -> FingerTree v a infixl 5

O(1). Add an element to the right end of a sequence. Mnemonic: a triangle with the single element at the pointy end.

(<|) :: Measured v a => a -> FingerTree v a -> FingerTree v a infixr 5

O(1). Add an element to the left end of a sequence. Mnemonic: a triangle with the single element at the pointy end.

fmap' :: (Measured v1 a1, Measured v2 a2) => (a1 -> a2) -> FingerTree v1 a1 -> FingerTree v2 a2

Like fmap, but with constraints on the element types.

data BadContext

Constructors

Shadowing

UnboundEx

UnboundUn

validateSegment :: Ord v => CtxSegment v loc -> Either BadContext (Context v loc)

Checks that a context segment forms a valid context. If so, then the proper context is yielded. Otherwise an indication of what is wrong with the segment is returned.

data SearchResult v a

A result of search, attempting to find a point where a predicate on splits of the sequence changes from False to True.

Since: fingertree-0.1.2.0

Constructors

Position !(FingerTree v a) a !(FingerTree v a)
A tree opened at a particular element: the prefix to the left, the element, and the suffix to the right.

OnLeft
A position to the left of the sequence, indicating that the predicate is True at both ends.

OnRight
A position to the right of the sequence, indicating that the predicate is False at both ends.

Nowhere
No position in the tree, returned if the predicate is True at the left end and False at the right end. This will not occur if the predicate in monotonic on the tree.

focusProblem :: Measured v e => (v -> Bool) -> FingerTree v e -> SearchResult v e

culprit :: Ord v => CtxSegment v loc -> BadContext -> (CtxSegment v loc, Element v loc)

info :: (Ord v, Measured (Info v loc) c) => c -> Info v loc

Return the aggregate Info associated to the context.

filter :: Var v => (Element v loc -> Bool) -> CtxSegment v loc -> CtxSegment v loc

partition :: Var v => (Element v loc -> Either (Element v loc) (Element v loc)) -> CtxSegment v loc -> (CtxSegment v loc, CtxSegment v loc)

apply :: (Var v, Measured (Info v loc) c) => c -> Type v loc -> Type v loc

Replace any existentials with their solution in the context

fromList :: Measured v a => [a] -> FingerTree v a

O(n). Create a sequence from a finite list of elements. The opposite operation toList is supplied by the Foldable instance.

mapMaybe :: (Element v loc -> Maybe b) -> CtxSegment v loc -> [b]

Applies a partial mapping to a context segment, resulting in a list.

single :: Var v => Element v loc -> CtxSegment v loc

split :: Ord v => v -> Context v loc -> Maybe (Context v loc, Element v loc, CtxSegment v loc)

splitSeg :: Ord v => v -> CtxSegment v loc -> Maybe (CtxSegment v loc, Element v loc, CtxSegment v loc)

surround :: Ord v => CtxSegment v loc -> Element v loc -> CtxSegment v loc -> CtxSegment v loc