Unison.Typechecker.TypeError

Documentation

data BooleanMismatch

Constructors

CondMismatch

AndMismatch

OrMismatch

GuardMismatch

data ExistentialMismatch

Constructors

IfBody

ListBody

CaseBody

data TypeError v loc

Constructors

Mismatch

Fields

foundType :: Type v loc

expectedType :: Type v loc

foundLeaf :: Type v loc

expectedLeaf :: Type v loc

mismatchSite :: Term v loc

note :: ErrorNote v loc

BooleanMismatch

Fields

getBooleanMismatch :: BooleanMismatch

mismatchSite :: Term v loc

foundType :: Type v loc

note :: ErrorNote v loc

ExistentialMismatch

Fields

getExistentialMismatch :: ExistentialMismatch

expectedType :: Type v loc

expectedLoc :: loc

foundType :: Type v loc

mismatchSite :: Term v loc

note :: ErrorNote v loc

FunctionApplication

Fields

f :: Term v loc

ft :: Type v loc

arg :: Term v loc

argNum :: Int

foundType :: Type v loc

expectedType :: Type v loc

leafs :: Maybe (Type v loc, Type v loc)

solvedVars :: [(v, Type v loc)]

note :: ErrorNote v loc

NotFunctionApplication

Fields

f :: Term v loc

ft :: Type v loc

note :: ErrorNote v loc

AbilityCheckFailure

Fields

ambient :: [Type v loc]

requested :: [Type v loc]

abilityCheckFailureSite :: loc

note :: ErrorNote v loc

AbilityEqFailure

Fields

lhs :: [Type v loc]

rhs :: [Type v loc]

tlhs :: Type v loc

trhs :: Type v loc

abilityCheckFailureSite :: loc

note :: ErrorNote v loc

AbilityEqFailureFromAp

Fields

lhs :: [Type v loc]

rhs :: [Type v loc]

tlhs :: Type v loc

trhs :: Type v loc

expectedSite :: Term v loc

mismatchSite :: Term v loc

note :: ErrorNote v loc

UnguardedLetRecCycle

Fields

cycle :: [v]

cycleLocs :: [loc]

note :: ErrorNote v loc

UnknownType

Fields

unknownTypeV :: v

typeSite :: loc

note :: ErrorNote v loc

UnknownTerm

Fields

unknownTermV :: v

termSite :: loc

suggestions :: [Suggestion v loc]

expectedType :: Type v loc

note :: ErrorNote v loc

DuplicateDefinitions

Fields

defns :: NonEmpty (v, [loc])

note :: ErrorNote v loc

UncoveredPatterns loc (NonEmpty (Pattern ()))

RedundantPattern loc

KindInferenceFailure (KindError v loc)

Other (ErrorNote v loc)

type RedundantTypeAnnotation = Bool

data TypeInfo v loc

Constructors

TopLevelComponent

Fields

definitions :: [(v, Type v loc, RedundantTypeAnnotation)]

type TypeNote v loc = Either (TypeError v loc) (TypeInfo v loc)

typeErrorFromNote :: (Ord loc, Show loc, Var v) => ErrorNote v loc -> TypeError v loc

typeInfoFromNote :: (Ord loc, Show loc, Var v) => InfoNote v loc -> Maybe (TypeInfo v loc)

allErrors :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

topLevelComponent :: InfoExtractor v a (TypeInfo v a)

redundantPattern :: ErrorExtractor v a (TypeError v a)

kindInferenceFailure :: ErrorExtractor v a (TypeError v a)

uncoveredPatterns :: ErrorExtractor v a (TypeError v a)

abilityCheckFailure :: ErrorExtractor v a (TypeError v a)

abilityEqFailure :: ErrorExtractor v a (TypeError v a)

duplicateDefinitions :: ErrorExtractor v a (TypeError v a)

unknownType :: ErrorExtractor v loc (TypeError v loc)

unknownTerm :: Var v => ErrorExtractor v loc (TypeError v loc)

generalMismatch :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

and :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

or :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

cond :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

matchGuard :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

unguardedCycle :: ErrorExtractor v loc (TypeError v loc)

booleanMismatch0 :: (Var v, Ord loc) => BooleanMismatch -> SubseqExtractor v loc () -> ErrorExtractor v loc (TypeError v loc)

helper function to support and or cond

existentialMismatch0 :: (Var v, Ord loc) => ExistentialMismatch -> SubseqExtractor v loc loc -> ErrorExtractor v loc (TypeError v loc)

ifBody :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

listBody :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

matchBody :: (Var v, Ord loc) => ErrorExtractor v loc (TypeError v loc)

applyingNonFunction :: Var v => ErrorExtractor v loc (TypeError v loc)

applyingFunction :: forall v loc. Var v => ErrorExtractor v loc (TypeError v loc)

Want to collect this info: The nth argument to $sel:f:Mismatch is $sel:foundType:Mismatch, but I was expecting $sel:expectedType:Mismatch.

30 | asdf asdf asdf

If you're curious $sel:f:Mismatch has type blah, where a was chosen as A b was chosen as B c was chosen as C (many colors / groups)

inSubtypes :: SubseqExtractor v loc (Type v loc, Type v loc, Maybe (Type v loc, Type v loc))