Unison.UnconflictedLocalDefnsView

Documentation

data UnconflictedLocalDefnsView

A view of a namespace's dsefinition (everything outside of lib) that is unconflicted: each name refers to one thing. The data contained within is all just different expressions of the same contents, for various use cases. The intention is to use laziness to avoid recomputing data structures whenever possible.

Constructors

UnconflictedLocalDefnsView

Fields

defns :: Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name)

nametree :: Nametree (DefnsF (Map NameSegment) Referent TypeReference)

names :: Names

empty :: UnconflictedLocalDefnsView

fromDefns :: DefnsF (Map Name) Referent TypeReference -> UnconflictedLocalDefnsView