Unison.UnisonFile.Names

Documentation

toNames :: Var v => UnisonFile v a -> Names

addNamesFromUnisonFile :: Var v => UnisonFile v a -> Names -> Names

typecheckedToNames :: Var v => TypecheckedUnisonFile v a -> Names

addNamesFromTypeCheckedUnisonFile :: Var v => TypecheckedUnisonFile v a -> Names -> Names

typecheckedUnisonFile0 :: Ord v => TypecheckedUnisonFile v a

bindNames :: Var v => Names -> UnisonFile v a -> ResolutionResult v a (UnisonFile v a)

variableCanonicalizer :: forall v. Var v => [v] -> Map v v

Given the set of fully-qualified variable names, this computes a Map from unique suffixes to the fully qualified name.

Example, given [foo.bar, qux.bar, baz.quaffle], this returns:

Map [ foo.bar -> foo.bar , qux.bar -> qux.bar , baz.quaffle -> baz.quaffle , quaffle -> baz.quaffle ]

This is used to replace variable references with their canonical fully qualified variables.

It's used below in environmentFor and also during the term resolution process.

environmentFor :: forall v a. Var v => Names -> Map v (DataDeclaration v a) -> Map v (EffectDeclaration v a) -> ResolutionResult v a (Either [Error v a] (Env v a))