Unison.UnisonFile.Summary

Documentation

data FileSummary

A file that parses might not always type-check, but often we just want to get as much information as we have available. This provides a type where we can summarize the information available in a Unison file.

If the file typechecked then all the Ref Ids and types will be filled in, otherwise they will be Nothing.

Constructors

FileSummary

Fields

dataDeclsBySymbol :: Map Symbol (Id, DataDeclaration Symbol Ann)

dataDeclsByReference :: Map Id (Map Symbol (DataDeclaration Symbol Ann))

effectDeclsBySymbol :: Map Symbol (Id, EffectDeclaration Symbol Ann)

effectDeclsByReference :: Map Id (Map Symbol (EffectDeclaration Symbol Ann))

termsBySymbol :: Map Symbol (Ann, Maybe Id, Term Symbol Ann, Maybe (Type Symbol Ann))

termsByReference :: Map (Maybe Id) (Map Symbol (Ann, Term Symbol Ann, Maybe (Type Symbol Ann)))

testWatchSummary :: [(Ann, Maybe Symbol, Maybe Id, Term Symbol Ann, Maybe (Type Symbol Ann))]

exprWatchSummary :: [(Ann, Maybe Symbol, Maybe Id, Term Symbol Ann, Maybe (Type Symbol Ann), Maybe WatchKind)]

fileNames :: Names

allWatches :: FileSummary -> [(Ann, Maybe Symbol, Maybe Id, Term Symbol Ann, Maybe (Type Symbol Ann), Maybe WatchKind)]

allTypeDecls :: FileSummary -> Map Symbol (Id, Either (EffectDeclaration Symbol Ann) (DataDeclaration Symbol Ann))

mkFileSummary :: Maybe (UnisonFile Symbol Ann) -> Maybe (TypecheckedUnisonFile Symbol Ann) -> Maybe FileSummary

Summarize the information available to us from the current state of the file. See FileSummary for more information.

fileDefLocations :: FileSummary -> Map Symbol (Set Ann)

Compute the location of user defined definitions within the file