Unison.Util.PinBoard

Description

A utility type for saving memory in the presence of many duplicate ByteStrings, etc. If you have data that may be a redundant duplicate, try pinning it to a pin board, and use the result of that operation instead.

Without a pin board:

x ───── "38dce848c8c829c62" y ───── "38dce848c8c829c62" z ───── "d2518f260535b927b"

With a pin board:

x ───── "38dce848c8c829c62" ┄┄┄┄┄┐ y ────────┘ board z ───── "d2518f260535b927b" ┄┄┄┄┄┘

... and after x is garbage collected:

"38dce848c8c829c62" ┄┄┄┄┄┐ y ────────┘ board z ───── "d2518f260535b927b" ┄┄┄┄┄┘

... and after y is garbage collected:

board z ───── "d2518f260535b927b" ┄┄┄┄┄┘

Documentation

data PinBoard a

A "pin board" is a place to pin values; semantically, it's a set, but differs in a few ways:

Pinned values aren't kept alive by the pin board, they might be garbage collected at any time.

If you try to pin a value that's already pinned (per its Eq instance), the pinned one will be returned instead.

It has a small API: just new and pin.

new :: MonadIO m => m (PinBoard a)

pin :: forall a m. (Eq a, Hashable a, MonadIO m) => PinBoard a -> a -> m a

For debugging

debugDump :: MonadIO m => (a -> Text) -> PinBoard a -> m ()

debugSize :: PinBoard a -> IO Int