Unison.Util.Text

Documentation

newtype Text

Constructors

Text (Rope Chunk)

data Chunk

Constructors

Chunk !Int !Text

empty :: Text

one :: Char -> Text

singleton :: Char -> Text

appendUnbalanced :: Text -> Text -> Text

threshold :: Int

replicate :: Int -> Text -> Text

toLazyText :: Text -> Text

chunkToText :: Chunk -> Text

chunk :: Text -> Chunk

take :: Int -> Text -> Text

drop :: Int -> Text -> Text

uncons :: Text -> Maybe (Char, Text)

unsnoc :: Text -> Maybe (Text, Char)

unconsChunk :: Text -> Maybe (Chunk, Text)

unsnocChunk :: Text -> Maybe (Text, Chunk)

at :: Int -> Text -> Maybe Char

size :: Text -> Int

reverse :: Text -> Text

toUppercase :: Text -> Text

toLowercase :: Text -> Text

fromUtf8 :: Bytes -> Either String Text

toUtf8 :: Text -> Bytes

fromText :: Text -> Text

pack :: String -> Text

toString :: Text -> String

unpack :: Text -> String

toText :: Text -> Text

indexOf :: Text -> Text -> Maybe Word64

ordinal :: IsString s => Int -> s

Return the ordinal representation of a number in English. A number ending with '1' must finish with st A number ending with '2' must finish with nd A number ending with '3' must finish with rd _except_ for 11, 12, and 13 which must finish with th

dropTextWhileMax :: (Char -> Bool) -> Int -> Text -> (Int, Text)

dropWhileMax :: (Char -> Bool) -> Int -> Text -> (Int, Text)