Unison.Codebase

Documentation

data Codebase m v a

Abstract interface to a user's codebase.

Terms

getTerm :: Codebase m v a -> Id -> m (Maybe (Term v a))

Get a user-defined term from the codebase.

Note that it is possible to call putTerm, then getTerm, and receive Nothing, per the semantics of putTerm.

unsafeGetTerm :: (HasCallStack, Monad m) => Codebase m v a -> Id -> m (Term v a)

Like getTerm, for when the term is known to exist in the codebase.

unsafeGetTermWithType :: (HasCallStack, Monad m) => Codebase m v a -> Id -> m (Term v a, Type v a)

Like unsafeGetTerm, but returns the type of the term, too.

getTypeOfTerm :: (Applicative m, Var v, BuiltinAnnotation a) => Codebase m v a -> Reference -> m (Maybe (Type v a))

Get the type of a term.

Note that it is possible to call putTerm, then getTypeOfTerm, and receive Nothing, per the semantics of putTerm.

unsafeGetTypeOfTermById :: (HasCallStack, Monad m) => Codebase m v a -> Id -> m (Type v a)

Like getTypeOfTerm, but for when the term is known to exist in the codebase.

isTerm :: (Applicative m, Var v, BuiltinAnnotation a) => Codebase m v a -> Reference -> m Bool

Check whether a reference is a term.

putTerm :: Codebase m v a -> Id -> Term v a -> Type v a -> m ()

Enqueue the put of a user-defined term (with its type) into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the term's (and its type's) references are stored as well.

Referents (sorta-termlike)

getTypeOfReferent :: (BuiltinAnnotation a, Var v, Monad m) => Codebase m v a -> Referent -> m (Maybe (Type v a))

Get the type of a referent.

Search

termsOfType :: (Var v, Functor m) => Codebase m v a -> Type v a -> m (Set Referent)

Get the set of terms-or-constructors that have the given type.

termsMentioningType :: (Var v, Functor m) => Codebase m v a -> Type v a -> m (Set Referent)

Get the set of terms-or-constructors mention the given type anywhere in their signature.

termReferencesByPrefix :: Codebase m v a -> ShortHash -> m (Set Id)

Get the set of user-defined terms whose hash matches the given prefix.

termReferentsByPrefix :: Codebase m v a -> ShortHash -> m (Set Id)

Get the set of user-defined terms-or-constructors whose hash matches the given prefix.

Type declarations

getTypeDeclaration :: Codebase m v a -> Id -> m (Maybe (Decl v a))

Get a type declaration.

Note that it is possible to call putTypeDeclaration, then getTypeDeclaration, and receive Nothing, per the semantics of putTypeDeclaration.

unsafeGetTypeDeclaration :: (HasCallStack, Monad m) => Codebase m v a -> Id -> m (Decl v a)

Like getTypeDeclaration, for when the type declaration is known to exist in the codebase.

putTypeDeclaration :: Codebase m v a -> Id -> Decl v a -> m ()

Enqueue the put of a type declaration into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the type declaration's references are stored as well.

typeReferencesByPrefix :: Codebase m v a -> ShortHash -> m (Set Id)

Get the set of type declarations whose hash matches the given prefix.

isType :: Applicative m => Codebase m v a -> Reference -> m Bool

Branches

branchExists :: Codebase m v a -> Hash -> m Bool

Check whether the given branch exists in the codebase.

getBranchForHash :: Monad m => Codebase m v a -> Hash -> m (Maybe (Branch m))

Get a branch from the codebase.

putBranch :: Codebase m v a -> Branch m -> m ()

Put a branch into the codebase, which includes its children, its patches, and the branch itself, if they don't already exist.

The terms and type declarations that a branch references must already exist in the codebase.

branchHashesByPrefix :: Codebase m v a -> ShortBranchHash -> m (Set Hash)

Get the set of branches whose hash matches the given prefix.

lca :: Monad m => Codebase m v a -> Branch m -> Branch m -> m (Maybe (Branch m))

Get the lowest common ancestor of two branches, i.e. the most recent branch that is an ancestor of both branches.

beforeImpl :: Codebase m v a -> Maybe (Hash -> Hash -> m Bool)

Root branch

getRootBranch :: Codebase m v a -> m (Either GetRootBranchError (Branch m))

Get the root branch.

data GetRootBranchError

Constructors

NoRootBranch

CouldntParseRootBranch String

CouldntLoadRootBranch Hash

isBlank :: Applicative m => Codebase m v a -> m Bool

Return whether the root branch is empty.

putRootBranch :: Codebase m v a -> Branch m -> m ()

Like putBranch, but also adjusts the root branch pointer afterwards.

rootBranchUpdates :: Codebase m v a -> m (IO (), IO (Set Hash))

Patches

patchExists :: Codebase m v a -> EditHash -> m Bool

Check whether the given patch exists in the codebase.

getPatch :: Codebase m v a -> EditHash -> m (Maybe Patch)

Get a patch from the codebase.

putPatch :: Codebase m v a -> EditHash -> Patch -> m ()

Put a patch into the codebase.

Note that putBranch may also put patches.

Watches

getWatch :: Codebase m v a -> WatchKind -> Id -> m (Maybe (Term v a))

getWatch k r returns watch result t that was previously put by putWatch k r t.

lookupWatchCache :: Monad m => Codebase m v a -> Id -> m (Maybe (Term v a))

Like getWatch, but first looks up the given reference as a regular watch, then as a test watch.

lookupWatchCache codebase ref =
  runMaybeT do
    MaybeT (getWatch codebase RegularWatch ref)
      | MaybeT (getWatch codebase TestWatch ref))

watches :: Codebase m v a -> WatchKind -> m [Id]

watches k returns all of the references r that were previously put by a putWatch k r t. t can be retrieved by getWatch k r.

putWatch :: Codebase m v a -> WatchKind -> Id -> Term v a -> m ()

putWatch k r t puts a watch of kind k, with hash-of-expression r and decompiled result t into the codebase.

For example, in the watch expression below, k is Regular, r is the hash of x, and t is 7.

> x = 3 + 4
  ⧩
  7

clearWatches :: Codebase m v a -> m ()

Delete all watches that were put by putWatch.

Reflog

getReflog :: Codebase m v a -> m [Entry Hash]

Get the entire reflog.

appendReflog :: Codebase m v a -> Text -> Branch m -> Branch m -> m ()

appendReflog reason before after appends a reflog entry.

FIXME: this could have type

appendReflog :: Reflog.Entry (Branch m) -> m ()

Unambiguous hash length

hashLength :: Codebase m v a -> m Int

The number of base32 characters needed to distinguish any two references in the codebase.

branchHashLength :: Codebase m v a -> m Int

The number of base32 characters needed to distinguish any two branch in the codebase.

Dependents

dependents :: Functor m => Codebase m v a -> Reference -> m (Set Reference)

Get the set of terms, type declarations, and builtin types that depend on the given term, type declaration, or builtin type.

Sync

Local sync

syncFromDirectory :: Codebase m v a -> CodebasePath -> SyncMode -> Branch m -> m ()

Copy a branch and all of its dependencies from the given codebase into this one.

syncToDirectory :: Codebase m v a -> CodebasePath -> SyncMode -> Branch m -> m ()

Copy a branch and all of its dependencies from this codebase into the given codebase.

Remote sync

viewRemoteBranch :: MonadIO m => Codebase m v a -> ReadRemoteNamespace -> m (Either GitError (m (), Branch m))

Pull a git branch and view it from the cache, without syncing into the local codebase.

importRemoteBranch :: forall m v a. MonadIO m => Codebase m v a -> ReadRemoteNamespace -> SyncMode -> m (Either GitError (Branch m))

Sync elements as needed from a remote codebase into the local one. If sbh is supplied, we try to load the specified branch hash; otherwise we try to load the root branch.

pushGitRootBranch :: Codebase m v a -> Branch m -> WriteRepo -> SyncMode -> m (Either GitError ())

Push the given branch to the given repo, and set it as the root branch.

Codebase path

getCodebaseDir :: MonadIO m => Maybe CodebasePath -> m CodebasePath

type CodebasePath = FilePath

A directory that contains a codebase.

type SyncToDir m = CodebasePath -> SyncMode -> Branch m -> m ()

Misc

addDefsToCodebase :: forall m v a. (Monad m, Var v, Show a) => Codebase m v a -> TypecheckedUnisonFile v a -> m ()

installUcmDependencies :: forall m. Monad m => Codebase m Symbol Ann -> m ()

Write all of UCM's dependencies (builtins types and an empty namespace) into the codebase

toCodeLookup :: Codebase m v a -> CodeLookup v m a

typeLookupForDependencies :: (Monad m, Var v, BuiltinAnnotation a) => Codebase m v a -> Set Reference -> m (TypeLookup v a)