Unison.Codebase

Documentation

data Codebase m v a

Abstract interface to a user's codebase.

One implementation is FileCodebase which uses the filesystem.

Constructors

Codebase

Fields

getTerm :: Id -> m (Maybe (Term v a))

getTypeOfTermImpl :: Id -> m (Maybe (Type v a))

getTypeDeclaration :: Id -> m (Maybe (Decl v a))

putTerm :: Id -> Term v a -> Type v a -> m ()

putTypeDeclaration :: Id -> Decl v a -> m ()

getRootBranch :: m (Either GetRootBranchError (Branch m))

putRootBranch :: Branch m -> m ()

rootBranchUpdates :: m (IO (), IO (Set Hash))

getBranchForHashImpl :: Hash -> m (Maybe (Branch m))

putBranch :: Branch m -> m ()

branchExists :: Hash -> m Bool

getPatch :: EditHash -> m (Maybe Patch)

putPatch :: EditHash -> Patch -> m ()

patchExists :: EditHash -> m Bool

dependentsImpl :: Reference -> m (Set Id)

syncFromDirectory :: CodebasePath -> SyncMode -> Branch m -> m ()

syncToDirectory :: CodebasePath -> SyncMode -> Branch m -> m ()

viewRemoteBranch' :: ReadRemoteNamespace -> m (Either GitError (m (), Branch m, CodebasePath))

pushGitRootBranch :: Branch m -> WriteRepo -> SyncMode -> m (Either GitError ())

watches :: WatchKind -> m [Id]

getWatch :: WatchKind -> Id -> m (Maybe (Term v a))

putWatch :: WatchKind -> Id -> Term v a -> m ()

clearWatches :: m ()

getReflog :: m [Entry Hash]

appendReflog :: Text -> Branch m -> Branch m -> m ()

termsOfTypeImpl :: Reference -> m (Set Id)

termsMentioningTypeImpl :: Reference -> m (Set Id)

hashLength :: m Int

termReferencesByPrefix :: ShortHash -> m (Set Id)

typeReferencesByPrefix :: ShortHash -> m (Set Id)

termReferentsByPrefix :: ShortHash -> m (Set Id)

branchHashLength :: m Int

branchHashesByPrefix :: ShortBranchHash -> m (Set Hash)

lcaImpl :: Maybe (Hash -> Hash -> m (Maybe Hash))

beforeImpl :: Maybe (Hash -> Hash -> m Bool)

type CodebasePath = FilePath

data GetRootBranchError

Constructors

NoRootBranch

CouldntParseRootBranch String

CouldntLoadRootBranch Hash

getBranchForHash :: Monad m => Codebase m v a -> Hash -> m (Maybe (Branch m))

getCodebaseDir :: MonadIO m => Maybe CodebasePath -> m CodebasePath

isBlank :: Applicative m => Codebase m v a -> m Bool

type SyncToDir m = CodebasePath -> SyncMode -> Branch m -> m ()

addDefsToCodebase :: forall m v a. (Monad m, Var v, Show a) => Codebase m v a -> TypecheckedUnisonFile v a -> m ()

installUcmDependencies :: forall m. Monad m => Codebase m Symbol Ann -> m ()

Write all of UCM's dependencies (builtins types and an empty namespace) into the codebase

getTypeOfTerm :: (Applicative m, Var v, BuiltinAnnotation a) => Codebase m v a -> Reference -> m (Maybe (Type v a))

getTypeOfReferent :: (BuiltinAnnotation a, Var v, Monad m) => Codebase m v a -> Referent -> m (Maybe (Type v a))

lca :: Monad m => Codebase m v a -> Branch m -> Branch m -> m (Maybe (Branch m))

lookupWatchCache :: Monad m => Codebase m v a -> Reference -> m (Maybe (Term v a))

toCodeLookup :: Codebase m v a -> CodeLookup v m a

typeLookupForDependencies :: (Monad m, Var v, BuiltinAnnotation a) => Codebase m v a -> Set Reference -> m (TypeLookup v a)

importRemoteBranch :: forall m v a. MonadIO m => Codebase m v a -> ReadRemoteNamespace -> SyncMode -> m (Either GitError (Branch m))

Sync elements as needed from a remote codebase into the local one. If sbh is supplied, we try to load the specified branch hash; otherwise we try to load the root branch.

viewRemoteBranch :: MonadIO m => Codebase m v a -> ReadRemoteNamespace -> m (Either GitError (m (), Branch m))

Pull a git branch and view it from the cache, without syncing into the local codebase.

termsOfType :: (Var v, Functor m) => Codebase m v a -> Type v a -> m (Set Referent)

termsMentioningType :: (Var v, Functor m) => Codebase m v a -> Type v a -> m (Set Referent)

dependents :: Functor m => Codebase m v a -> Reference -> m (Set Reference)

The dependents of a builtin type includes the set of builtin terms which mention that type.

isTerm :: (Applicative m, Var v, BuiltinAnnotation a) => Codebase m v a -> Reference -> m Bool

isType :: Applicative m => Codebase m v a -> Reference -> m Bool