Unison.Codebase

Documentation

data Codebase m v a

Abstract interface to a user's codebase.

Terms

$sel:getTerm:Codebase :: Codebase m v a -> Id -> Transaction (Maybe (Term v a))

Get a user-defined term from the codebase.

Note that it is possible to call $sel:putTerm:Codebase, then $sel:getTerm:Codebase, and receive Nothing, per the semantics of $sel:putTerm:Codebase.

unsafeGetTerm :: HasCallStack => Codebase m v a -> Id -> Transaction (Term v a)

Like $sel:getTerm:Codebase, for when the term is known to exist in the codebase.

unsafeGetTermWithType :: HasCallStack => Codebase m v a -> Id -> Transaction (Term v a, Type v a)

Like unsafeGetTerm, but returns the type of the term, too.

$sel:getTermComponentWithTypes:Codebase :: Codebase m v a -> Hash -> Transaction (Maybe [(Term v a, Type v a)])

unsafeGetTermComponent :: HasCallStack => Codebase m v a -> Hash -> Transaction [(Term v a, Type v a)]

Like $sel:getTermComponentWithTypes:Codebase, for when the term component is known to exist in the codebase.

getTypeOfTerm :: BuiltinAnnotation a => Codebase m Symbol a -> Reference -> Transaction (Maybe (Type Symbol a))

Get the type of a term.

Note that it is possible to call $sel:putTerm:Codebase, then getTypeOfTerm, and receive Nothing, per the semantics of $sel:putTerm:Codebase.

$sel:getDeclType:Codebase :: Codebase m v a -> Reference -> Transaction ConstructorType

Get the type of a given decl.

unsafeGetTypeOfTermById :: HasCallStack => Codebase m v a -> Id -> Transaction (Type v a)

Like getTypeOfTerm, but for when the term is known to exist in the codebase.

isTerm :: BuiltinAnnotation a => Codebase m Symbol a -> Reference -> Transaction Bool

Check whether a reference is a term.

$sel:putTerm:Codebase :: Codebase m v a -> Id -> Term v a -> Type v a -> Transaction ()

Enqueue the put of a user-defined term (with its type) into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the term's (and its type's) references are stored as well.

$sel:putTermComponent:Codebase :: Codebase m v a -> Hash -> [(Term v a, Type v a)] -> Transaction ()

Referents (sorta-termlike)

getTypeOfReferent :: BuiltinAnnotation a => Codebase m Symbol a -> Referent -> Transaction (Maybe (Type Symbol a))

Get the type of a referent.

Search

termsOfType :: Var v => Codebase m v a -> Type v a -> Transaction (Set Referent)

Get the set of terms-or-constructors that have the given type.

filterTermsByReferenceIdHavingType :: Var v => Codebase m v a -> Type v a -> Set TermReferenceId -> Transaction (Set TermReferenceId)

filterTermsByReferentHavingType :: Var v => Codebase m v a -> Type v a -> Set Referent -> Transaction (Set Referent)

termsMentioningType :: Var v => Codebase m v a -> Type v a -> Transaction (Set Referent)

Get the set of terms-or-constructors mention the given type anywhere in their signature.

termReferencesByPrefix :: ShortHash -> Transaction (Set Id)

$sel:termReferentsByPrefix:Codebase :: Codebase m v a -> ShortHash -> Transaction (Set Id)

Get the set of user-defined terms-or-constructors whose hash matches the given prefix.

Type declarations

$sel:getTypeDeclaration:Codebase :: Codebase m v a -> Id -> Transaction (Maybe (Decl v a))

Get a type declaration.

Note that it is possible to call $sel:putTypeDeclaration:Codebase, then $sel:getTypeDeclaration:Codebase, and receive Nothing, per the semantics of $sel:putTypeDeclaration:Codebase.

unsafeGetTypeDeclaration :: HasCallStack => Codebase m v a -> Id -> Transaction (Decl v a)

Like $sel:getTypeDeclaration:Codebase, for when the type declaration is known to exist in the codebase.

getDeclComponent :: Hash -> Transaction (Maybe [Decl Symbol Ann])

$sel:putTypeDeclaration:Codebase :: Codebase m v a -> Id -> Decl v a -> Transaction ()

Enqueue the put of a type declaration into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the type declaration's references are stored as well.

$sel:putTypeDeclarationComponent:Codebase :: Codebase m v a -> Hash -> [Decl v a] -> Transaction ()

typeReferencesByPrefix :: ShortHash -> Transaction (Set Id)

Get the set of type declarations whose hash matches the given prefix.

isType :: Codebase m v a -> Reference -> Transaction Bool

Branches

branchExists :: CausalHash -> Transaction Bool

Check whether the given branch exists in the codebase.

$sel:getBranchForHash:Codebase :: Codebase m v a -> CausalHash -> m (Maybe (Branch m))

expectBranchForHash :: Monad m => Codebase m v a -> CausalHash -> m (Branch m)

Like $sel:getBranchForHash:Codebase, but for when the hash is known to be in the codebase.

$sel:putBranch:Codebase :: Codebase m v a -> Branch m -> m ()

Put a branch into the codebase, which includes its children, its patches, and the branch itself, if they don't already exist.

The terms and type declarations that a branch references must already exist in the codebase.

causalHashesByPrefix :: ShortCausalHash -> Transaction (Set CausalHash)

Get the set of branches whose hash matches the given prefix.

lca :: MonadIO m => Codebase m v a -> Branch m -> Branch m -> m (Maybe (Branch m))

Get the lowest common ancestor of two branches, i.e. the most recent branch that is an ancestor of both branches.

before :: CausalHash -> CausalHash -> Transaction Bool

getShallowBranchAtPath :: Path -> Maybe (Branch Transaction) -> Transaction (Branch Transaction)

Recursively descend into causals following the given path, Use the root causal if none is provided.

getShallowCausalAtPath :: Path -> Maybe (CausalBranch Transaction) -> Transaction (CausalBranch Transaction)

Recursively descend into causals following the given path, Use the root causal if none is provided.

getBranchAtPath :: MonadIO m => Codebase m v a -> Absolute -> m (Branch m)

Get a v1 branch from the root following the given path.

expectCausalBranchByCausalHash :: CausalHash -> Transaction (CausalBranch Transaction)

getShallowCausalFromRoot :: Maybe CausalHash -> Path -> Transaction (CausalBranch Transaction)

getShallowRootBranch :: Transaction (Branch Transaction)

Get the shallow representation of the root branches without loading the children or history.

getShallowRootCausal :: Transaction (CausalBranch Transaction)

Get the shallow representation of the root branches without loading the children or history.

Root branch

$sel:getRootBranch:Codebase :: Codebase m v a -> m (Branch m)

Get the root branch.

getRootBranchExists :: Transaction Bool

Get whether the root branch exists.

expectRootCausalHash :: Transaction CausalHash

$sel:putRootBranch:Codebase :: Codebase m v a -> Text -> Branch m -> m ()

Like $sel:putBranch:Codebase, but also adjusts the root branch pointer afterwards.

namesAtPath :: BranchHash -> Path -> Transaction Names

Construct a ScopedNames which can produce names which are relative to the provided Path.

NOTE: this method requires an up-to-date name lookup index

Patches

patchExists :: PatchHash -> Transaction Bool

Check whether the given patch exists in the codebase.

getPatch :: PatchHash -> Transaction (Maybe Patch)

putPatch :: PatchHash -> Patch -> Transaction ()

Put a patch into the codebase.

Note that putBranch may also put patches.

Watches

$sel:getWatch:Codebase :: Codebase m v a -> WatchKind -> Id -> Transaction (Maybe (Term v a))

getWatch k r returns watch result t that was previously put by putWatch k r t.

lookupWatchCache :: Codebase m v a -> Id -> Transaction (Maybe (Term v a))

Like $sel:getWatch:Codebase, but first looks up the given reference as a regular watch, then as a test watch.

lookupWatchCache codebase ref =
  runMaybeT do
    MaybeT (getWatch codebase RegularWatch ref)
      | MaybeT (getWatch codebase TestWatch ref))

watches :: WatchKind -> Transaction [Id]

watches k returns all of the references r that were previously put by a putWatch k r t. t can be retrieved by getWatch k r.

putWatch :: WatchKind -> Id -> Term Symbol Ann -> Transaction ()

putWatch k r t puts a watch of kind k, with hash-of-expression r and decompiled result t into the codebase.

For example, in the watch expression below, k is Regular, r is the hash of x, and t is 7.

> x = 3 + 4
  ⧩
  7

clearWatches :: Transaction ()

Delete all watches that were put by putWatch.

Reflog

getReflog :: Int -> Transaction [Entry CausalHash Text]

Gets the specified number of reflog entries in chronological order, most recent first.

Unambiguous hash length

hashLength :: Transaction Int

The number of base32 characters needed to distinguish any two references in the codebase.

branchHashLength :: Transaction Int

The number of base32 characters needed to distinguish any two branch in the codebase.

Dependents

dependents :: DependentsSelector -> Reference -> Transaction (Set Reference)

Get the set of terms, type declarations, and builtin types that depend on the given term, type declaration, or builtin type.

dependentsOfComponent :: Hash -> Transaction (Set Reference)

Sync

Local sync

$sel:syncFromDirectory:Codebase :: Codebase m v a -> CodebasePath -> SyncMode -> Branch m -> m ()

Copy a branch and all of its dependencies from the given codebase into this one.

$sel:syncToDirectory:Codebase :: Codebase m v a -> CodebasePath -> SyncMode -> Branch m -> m ()

Copy a branch and all of its dependencies from this codebase into the given codebase.

Remote sync

viewRemoteBranch :: MonadIO m => Codebase m v a -> ReadGitRemoteNamespace -> GitBranchBehavior -> (Branch m -> m r) -> m (Either GitError r)

Pull a git branch and view it from the cache, without syncing into the local codebase.

importRemoteBranch :: forall m v a. MonadUnliftIO m => Codebase m v a -> ReadGitRemoteNamespace -> SyncMode -> Preprocessing m -> m (Either GitError (Branch m))

Sync elements as needed from a remote git codebase into the local one. If sch is supplied, we try to load the specified branch hash; otherwise we try to load the root branch.

data Preprocessing m

An optional preprocessing step to run on branches before they're imported into the local codebase.

Constructors

Unmodified

Preprocessed (Branch m -> m (Branch m))

$sel:pushGitBranch:Codebase :: Codebase m v a -> forall e. WriteGitRepo -> PushGitBranchOpts -> (Branch m -> m (Either e (Branch m))) -> m (Either GitError (Either e (Branch m)))

Push the given branch to the given repo, and optionally set it as the root branch.

data PushGitBranchOpts

Constructors

PushGitBranchOpts

Fields

behavior :: GitPushBehavior

syncMode :: SyncMode

Codebase path

getCodebaseDir :: MonadIO m => Maybe CodebasePath -> m CodebasePath

type CodebasePath = FilePath

A directory that contains a codebase.

type SyncToDir m = CodebasePath -> SyncMode -> Branch m -> m ()

Direct codebase access

runTransaction :: MonadIO m => Codebase m v a -> Transaction b -> m b

Run a transaction on a codebase.

runTransactionWithRollback :: MonadIO m => Codebase m v a -> ((forall void. b -> Transaction void) -> Transaction b) -> m b

$sel:withConnection:Codebase :: Codebase m v a -> forall x. (Connection -> m x) -> m x

Acquire a new connection to the same underlying database file this codebase object connects to.

$sel:withConnectionIO:Codebase :: Codebase m v a -> forall x. (Connection -> IO x) -> IO x

Acquire a new connection to the same underlying database file this codebase object connects to.

Misc (organize these better)

addDefsToCodebase :: forall m v a. (Var v, Show a) => Codebase m v a -> TypecheckedUnisonFile v a -> Transaction ()

componentReferencesForReference :: Reference -> Transaction (Set Reference)

installUcmDependencies :: Codebase m Symbol Ann -> Transaction ()

Write all of UCM's dependencies (builtins types and an empty namespace) into the codebase

toCodeLookup :: MonadIO m => Codebase m Symbol Ann -> CodeLookup Symbol m Ann

typeLookupForDependencies :: Codebase IO Symbol Ann -> Set Reference -> Transaction (TypeLookup Symbol Ann)

Make a TypeLookup that is suitable for looking up information about all of the given type-or-term references, and all of their type dependencies, including builtins.

unsafeGetComponentLength :: HasCallStack => Hash -> Transaction CycleSize