Unison.Codebase.Branch.Type

Documentation

newtype Branch m

A node in the Unison namespace hierarchy along with its history.

Constructors

Branch

Fields

_history :: UnwrappedBranch m

type UnwrappedBranch m = Causal m Raw (Branch0 m)

type Hash = RawHash Raw

type EditHash = Hash

type Star r n = Star r n

head :: Branch m -> Branch0 m

headHash :: Branch m -> Hash

data Branch0 m

A node in the Unison namespace hierarchy.

_terms and _types are the declarations at this level. _children are the nodes one level below us. _edits are the Patchs stored at this node in the code.

The deep* fields are derived from the four above.

Constructors

Branch0

Fields

_terms :: Star Referent NameSegment

_types :: Star Reference NameSegment

_children :: Map NameSegment (Branch m)

Note the Branch here, not Branch0. Every level in the tree has a history.

_edits :: Map NameSegment (EditHash, m Patch)

deepTerms :: Relation Referent Name

deepTypes :: Relation Reference Name

deepTermMetadata :: R4 Referent Name

deepTypeMetadata :: R4 Reference Name

deepPaths :: Set Path

deepEdits :: Map Name EditHash

history :: Iso' (Branch m) (UnwrappedBranch m)

edits :: Lens' (Branch0 m) (Map NameSegment (EditHash, m Patch))