Unison.Codebase.Branch.Type

Documentation

type NamespaceHash m = HashFor (Branch0 m)

A Hash for a namespace itself, it doesn't incorporate any history.

head :: Branch m -> Branch0 m

headHash :: Branch m -> CausalHash

namespaceHash :: Branch m -> NamespaceHash m

newtype Branch m

A node in the Unison namespace hierarchy along with its history.

Constructors

Branch

Fields

_history :: UnwrappedBranch m

data Branch0 m

A node in the Unison namespace hierarchy.

$sel:_terms:Branch0 and $sel:_types:Branch0 are the declarations at this level. $sel:_children:Branch0 are the nodes one level below us. $sel:_edits:Branch0 are the Patchs stored at this node in the code.

The remaining fields are derived from the four above. Please don't set them manually; use Branch.empty0 or Branch.branch0 to construct them.

Constructors

Branch0

Fields

_terms :: Star Referent NameSegment

_types :: Star Reference NameSegment

_children :: Map NameSegment (Branch m)

Note the Branch here, not Branch0. Every level in the tree has a history.

_edits :: Map NameSegment (PatchHash, m Patch)

isEmpty0 :: Bool

True if a branch and its children have no definitions or edits in them. (Computed recursively, and small enough to justify storing here to avoid computing more than once.)

deepTerms :: Relation Referent Name

deepTypes :: Relation Reference Name

deepPaths :: Set Path

deepEdits :: Map Name PatchHash

history :: Iso' (Branch m) (UnwrappedBranch m)

edits :: Lens' (Branch0 m) (Map NameSegment (PatchHash, m Patch))

type Star r n = Star r n

type UnwrappedBranch m = Causal m (Branch0 m)