Unison.Codebase.Init

Documentation

data Init m v a

Constructors

Init

Fields

withOpenCodebase :: forall r. DebugName -> CodebasePath -> CodebaseLockOption -> MigrationStrategy -> (Codebase m v a -> m r) -> m (Either OpenCodebaseError r)

open an existing codebase

withCreatedCodebase :: forall r. DebugName -> CodebasePath -> CodebaseLockOption -> (Codebase m v a -> m r) -> m (Either CreateCodebaseError r)

create a new codebase

codebasePath :: CodebasePath -> CodebasePath

given a codebase root, and given that the codebase root may have other junk in it, give the path to the "actual" files; e.g. what a forked transcript should clone.

type DebugName = String

data InitError

An error that occurred while initializing a codebase.

Constructors

FoundV1Codebase

InitErrorOpen OpenCodebaseError

CouldntCreateCodebase Pretty

data CodebaseInitOptions

Constructors

Home CodebasePath

Specified SpecifiedCodebase

data CodebaseLockOption

Constructors

DoLock

DontLock

data InitResult

Constructors

OpenedCodebase

CreatedCodebase

data SpecifiedCodebase

Constructors

CreateWhenMissing CodebasePath

DontCreateWhenMissing CodebasePath

data MigrationStrategy

Constructors

MigrateAutomatically BackupStrategy VacuumStrategy
Perform a migration immediately if one is required.

MigrateAfterPrompt BackupStrategy VacuumStrategy
Prompt the user that a migration is about to occur, continue after acknownledgment

DontMigrate
Triggers an OpenCodebaseRequiresMigration error instead of migrating

data BackupStrategy

Constructors

Backup

NoBackup

data VacuumStrategy

Constructors

Vacuum

NoVacuum

type Pretty = Pretty ColorText

createCodebase :: MonadIO m => Init m v a -> DebugName -> CodebasePath -> CodebaseLockOption -> (Codebase m v a -> m r) -> m (Either Pretty r)

initCodebaseAndExit :: MonadIO m => Init m Symbol Ann -> Verbosity -> DebugName -> Maybe CodebasePath -> CodebaseLockOption -> m ()

try to init a codebase where none exists and then exit regardless (i.e. `ucm --codebase dir init`)

withOpenOrCreateCodebase :: MonadIO m => Init m v a -> DebugName -> CodebaseInitOptions -> CodebaseLockOption -> MigrationStrategy -> ((InitResult, CodebasePath, Codebase m v a) -> m r) -> m (Either (CodebasePath, InitError) r)

withNewUcmCodebaseOrExit :: MonadIO m => Init m Symbol Ann -> Verbosity -> DebugName -> CodebasePath -> CodebaseLockOption -> (Codebase m Symbol Ann -> m r) -> m r

withTemporaryUcmCodebase :: MonadUnliftIO m => Init m Symbol Ann -> Verbosity -> DebugName -> CodebaseLockOption -> ((CodebasePath, Codebase m Symbol Ann) -> m r) -> m r