Unison.Codebase.Path

Documentation

newtype Path

A Path is an internal structure representing some namespace in the codebase.

Foo.Bar.baz becomes [Foo, Bar, "baz"].

NB: This shouldn’t be exposed outside of this module (preferPath', Absolute, or Relative), but it’s currently used pretty widely. Such usage should be replaced when encountered.

Constructors

Path

Fields

toSeq :: Seq NameSegment

newtype Path'

A namespace that may be either absolute or relative, This is the most general type that should be used.

Constructors

Path'

Fields

unPath' :: Either Absolute Relative

newtype Absolute

An absolute from the current project root

Constructors

Absolute

Fields

unabsolute :: Path

pattern AbsolutePath' :: Absolute -> Path'

absPath_ :: Lens' Absolute Path

newtype Relative

A namespace path that doesn’t necessarily start from the root. Typically refers to a path from the current namespace.

Constructors

Relative

Fields

unrelative :: Path

relPath_ :: Lens' Relative Path

pattern RelativePath' :: Relative -> Path'

class Resolve l r o where

Methods

resolve :: l -> r -> o

pattern Empty :: Path

pattern (:<) :: Cons b b a a => a -> b -> b infixr 5

pattern (:>) :: Snoc a a b b => a -> b -> a infixl 5

singleton :: NameSegment -> Path

uncons :: Path -> Maybe (NameSegment, Path)

empty :: Path

isAbsolute :: Path' -> Bool

isRelative :: Path' -> Bool

absoluteEmpty :: Absolute

absoluteEmpty' :: Path'

relativeEmpty :: Relative

relativeEmpty' :: Path'

currentPath :: Path'

prefix :: Path' -> Relative -> Path'

This always prefixes, since the secend argument can never be Absolute.

prefixAbs :: Absolute -> Relative -> Absolute

prefixRel :: Relative -> Relative -> Relative

maybePrefix :: Path' -> Path' -> Maybe Path'

Returns Nothing if the second argument is absolute. A common pattern is fromMaybe path $ maybePrefix prefix path to use the unmodified path in that case.

unprefix :: Absolute -> Path' -> Path

examples: unprefix .foo.bar .blah == .blah (absolute paths left alone) unprefix .foo.bar id == id (relative paths starting w/ nonmatching prefix left alone) unprefix .foo.bar foo.bar.baz == baz (relative paths w/ common prefix get stripped)

maybePrefixName :: Path' -> Name -> Maybe Name

Returns Nothing if the second argument is absolute. A common pattern is fromMaybe name $ maybePrefixName prefix name to use the unmodified path in that case.

prefixNameIfRel :: Path' -> Name -> Name

unprefixName :: Absolute -> Name -> Maybe Name

Remove a path prefix from a name. Returns Nothing if there are no remaining segments to construct the name from.

>>> unprefixName (Absolute $ fromList ["base", "List"]) (Name.unsafeFromText "base.List.map")
Just (Name Relative (NameSegment {toText = "map"} :| []))

type HQSplit = (Path, HQSegment)

type HQSplitAbsolute = (Absolute, HQSegment)

type AbsSplit = (Absolute, NameSegment)

type Split = (Path, NameSegment)

type Split' = (Path', NameSegment)

type HQSplit' = (Path', HQSegment)

ancestors :: Absolute -> Seq Absolute

utilities

longestPathPrefix :: Path -> Path -> (Path, Path, Path)

Finds the longest shared path prefix of two paths. Returns (shared prefix, path to first location from shared prefix, path to second location from shared prefix)

>>> longestPathPrefix ("a" :< "b" :< "x" :< Empty) ("a" :< "b" :< "c" :< Empty)
(a.b,x,c)

>>> longestPathPrefix Empty ("a" :< "b" :< "c" :< Empty)
(,,a.b.c)

tests

isCurrentPath :: Path' -> Bool

isRoot :: Absolute -> Bool

isRoot' :: Path' -> Bool

conversions

absoluteToPath' :: Absolute -> Path'

fromList :: [NameSegment] -> Path

fromName :: Name -> Path

fromName' :: Name -> Path'

fromPath' :: Path' -> Path

unsafeParseText :: Text -> Path

unsafeParseText' :: Text -> Path'

Construct a Path' from a text

>>> fromText' "a.b.c"
a.b.c

>>> fromText' ".a.b.c"
.a.b.c

>>> show $ fromText' ""
""

toAbsoluteSplit :: Absolute -> (Path', a) -> (Absolute, a)

toSplit' :: Path' -> Maybe (Path', NameSegment)

toList :: Path -> [NameSegment]

toName :: Path -> Maybe Name

toName' :: Path' -> Maybe Name

Convert a Path' to a Name

toText :: Path -> Text

Note: This treats the path as relative.

toText' :: Path' -> Text

absToText :: Absolute -> Text

relToText :: Relative -> Text

unsplit :: Split -> Path

unsplit' :: Split' -> Path'

unsplitAbsolute :: (Absolute, NameSegment) -> Absolute

nameFromHQSplit :: HQSplit -> HashQualified Name

nameFromHQSplit' :: HQSplit' -> HashQualified Name

nameFromSplit' :: Split' -> Name

splitFromName :: Name -> Split

>>> splitFromName "a.b.c"
(a.b,c)

>>> splitFromName "foo"
(,foo)

splitFromName' :: Name -> Split'

hqSplitFromName' :: Name -> HQSplit'

things that could be replaced with Cons instances

cons :: NameSegment -> Path -> Path

things that could be replaced with Snoc instances

snoc :: Path -> NameSegment -> Path

unsnoc :: Path -> Maybe (Path, NameSegment)