Unison.Codebase.SqliteCodebase.Migrations

Documentation

type Migration m a v = Connection -> Codebase m v a -> m (Either MigrationError ())

currentSchemaVersion :: SchemaVersion

The highest schema that this ucm knows how to migrate to.

migrations :: forall m v a. (MonadUnliftIO m, Var v) => Map SchemaVersion (Migration m a v)

Mapping from schema version to the migration required to get there. Each migration may only be run on a schema of its immediate predecessor, E.g. The migration at index 2 must be run on a codebase at version 1.

ensureCodebaseIsUpToDate :: (MonadUnliftIO m, Var v) => LocalOrRemote -> CodebasePath -> Connection -> Codebase m v a -> m (Either OpenCodebaseError ())

Migrates a codebase up to the most recent version known to ucm. This is a No-op if it's up to date Returns an error if the schema version is newer than this ucm knows about.

backupCodebase :: CodebasePath -> MonadIO m => m ()

Copy the sqlite database to a new file with a unique name based on current time.