Unison.Codebase.Type

Documentation

data Codebase m v a

Abstract interface to a user's codebase.

Constructors

Codebase

Fields

getTerm :: Id -> m (Maybe (Term v a))

Get a user-defined term from the codebase.

Note that it is possible to call putTerm, then getTerm, and receive Nothing, per the semantics of putTerm.

getTypeOfTermImpl :: Id -> m (Maybe (Type v a))

Get the type of a user-defined term.

Note that it is possible to call putTerm, then getTypeOfTermImpl, and receive Nothing, per the semantics of putTerm.

getTypeDeclaration :: Id -> m (Maybe (Decl v a))

Get a type declaration.

Note that it is possible to call putTypeDeclaration, then getTypeDeclaration, and receive Nothing, per the semantics of putTypeDeclaration.

putTerm :: Id -> Term v a -> Type v a -> m ()

Enqueue the put of a user-defined term (with its type) into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the term's (and its type's) references are stored as well.

putTypeDeclaration :: Id -> Decl v a -> m ()

Enqueue the put of a type declaration into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the type declaration's references are stored as well.

getRootBranch :: m (Either GetRootBranchError (Branch m))

Get the root branch.

getRootBranchExists :: m Bool

Get whether the root branch exists.

putRootBranch :: Branch m -> m ()

Like putBranch, but also adjusts the root branch pointer afterwards.

rootBranchUpdates :: m (IO (), IO (Set Hash))

getBranchForHashImpl :: Hash -> m (Maybe (Branch m))

putBranch :: Branch m -> m ()

Put a branch into the codebase, which includes its children, its patches, and the branch itself, if they don't already exist.

The terms and type declarations that a branch references must already exist in the codebase.

branchExists :: Hash -> m Bool

Check whether the given branch exists in the codebase.

getPatch :: EditHash -> m (Maybe Patch)

Get a patch from the codebase.

putPatch :: EditHash -> Patch -> m ()

Put a patch into the codebase.

Note that putBranch may also put patches.

patchExists :: EditHash -> m Bool

Check whether the given patch exists in the codebase.

dependentsImpl :: Reference -> m (Set Id)

Get the set of user-defined terms and type declarations that depend on the given term, type declaration, or builtin type.

syncFromDirectory :: CodebasePath -> SyncMode -> Branch m -> m ()

Copy a branch and all of its dependencies from the given codebase into this one.

syncToDirectory :: CodebasePath -> SyncMode -> Branch m -> m ()

Copy a branch and all of its dependencies from this codebase into the given codebase.

viewRemoteBranch' :: forall r. ReadRemoteNamespace -> ((Branch m, CodebasePath) -> m r) -> m (Either GitError r)

pushGitBranch :: Branch m -> WriteRepo -> PushGitBranchOpts -> m (Either GitError ())

Push the given branch to the given repo, and optionally set it as the root branch.

watches :: WatchKind -> m [Id]

watches k returns all of the references r that were previously put by a putWatch k r t. t can be retrieved by getWatch k r.

getWatch :: WatchKind -> Id -> m (Maybe (Term v a))

getWatch k r returns watch result t that was previously put by putWatch k r t.

putWatch :: WatchKind -> Id -> Term v a -> m ()

putWatch k r t puts a watch of kind k, with hash-of-expression r and decompiled result t into the codebase.

For example, in the watch expression below, k is Regular, r is the hash of x, and t is 7.

> x = 3 + 4
  ⧩
  7

clearWatches :: m ()

Delete all watches that were put by putWatch.

getReflog :: m [Entry Hash]

Get the entire reflog.

appendReflog :: Text -> Branch m -> Branch m -> m ()

appendReflog reason before after appends a reflog entry.

FIXME: this could have type

appendReflog :: Reflog.Entry (Branch m) -> m ()

termsOfTypeImpl :: Reference -> m (Set Id)

Get the set of user-defined terms-or-constructors that have the given type.

termsMentioningTypeImpl :: Reference -> m (Set Id)

Get the set of user-defined terms-or-constructors mention the given type anywhere in their signature.

hashLength :: m Int

The number of base32 characters needed to distinguish any two references in the codebase.

termReferencesByPrefix :: ShortHash -> m (Set Id)

Get the set of user-defined terms whose hash matches the given prefix.

typeReferencesByPrefix :: ShortHash -> m (Set Id)

Get the set of type declarations whose hash matches the given prefix.

termReferentsByPrefix :: ShortHash -> m (Set Id)

Get the set of user-defined terms-or-constructors whose hash matches the given prefix.

branchHashLength :: m Int

The number of base32 characters needed to distinguish any two branch in the codebase.

branchHashesByPrefix :: ShortBranchHash -> m (Set Hash)

Get the set of branches whose hash matches the given prefix.

lcaImpl :: Maybe (Hash -> Hash -> m (Maybe Hash))

beforeImpl :: Maybe (Hash -> Hash -> m Bool)

type CodebasePath = FilePath

A directory that contains a codebase.

data PushGitBranchOpts

Constructors

PushGitBranchOpts

Fields

setRoot :: Bool

Set the branch as root?

syncMode :: SyncMode

data GitError

Constructors

GitProtocolError GitProtocolError

GitCodebaseError (GitCodebaseError Hash)

GitSqliteCodebaseError GitSqliteCodebaseError

data GetRootBranchError

Constructors

NoRootBranch

CouldntParseRootBranch String

CouldntLoadRootBranch Hash

type SyncToDir m = CodebasePath -> SyncMode -> Branch m -> m ()