Unison.Codebase.Type

Documentation

data Codebase m v a

Abstract interface to a user's codebase.

Constructors

Codebase

Fields

getTerm :: TermReferenceId -> Transaction (Maybe (Term v a))

Get a user-defined term from the codebase.

Note that it is possible to call $sel:putTerm:Codebase, then $sel:getTerm:Codebase, and receive Nothing, per the semantics of $sel:putTerm:Codebase.

getTypeOfTermImpl :: TermReferenceId -> Transaction (Maybe (Type v a))

Get the type of a user-defined term.

Note that it is possible to call $sel:putTerm:Codebase, then $sel:getTypeOfTermImpl:Codebase, and receive Nothing, per the semantics of $sel:putTerm:Codebase.

getTypeDeclaration :: TypeReferenceId -> Transaction (Maybe (Decl v a))

Get a type declaration.

Note that it is possible to call $sel:putTypeDeclaration:Codebase, then $sel:getTypeDeclaration:Codebase, and receive Nothing, per the semantics of $sel:putTypeDeclaration:Codebase.

getDeclType :: TypeReference -> Transaction ConstructorType

Get the type of a given decl.

putTerm :: TermReferenceId -> Term v a -> Type v a -> Transaction ()

Enqueue the put of a user-defined term (with its type) into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the term's (and its type's) references are stored as well.

putTermComponent :: Hash -> [(Term v a, Type v a)] -> Transaction ()

putTypeDeclaration :: TypeReferenceId -> Decl v a -> Transaction ()

Enqueue the put of a type declaration into the codebase, if it doesn't already exist. The implementation may choose to delay the put until all of the type declaration's references are stored as well.

putTypeDeclarationComponent :: Hash -> [Decl v a] -> Transaction ()

getTermComponentWithTypes :: Hash -> Transaction (Maybe [(Term v a, Type v a)])

getBranchForHash :: CausalHash -> m (Maybe (Branch m))

putBranch :: Branch m -> m ()

Put a branch into the codebase, which includes its children, its patches, and the branch itself, if they don't already exist.

The terms and type declarations that a branch references must already exist in the codebase.

getWatch :: WatchKind -> TermReferenceId -> Transaction (Maybe (Term v a))

getWatch k r returns watch result t that was previously put by putWatch k r t.

termsOfTypeImpl :: Reference -> Transaction (Set Id)

Get the set of user-defined terms-or-constructors that have the given type.

termsMentioningTypeImpl :: Reference -> Transaction (Set Id)

Get the set of user-defined terms-or-constructors mention the given type anywhere in their signature.

filterTermsByReferenceIdHavingTypeImpl :: TypeReference -> Set Id -> Transaction (Set Id)

Return the subset of the given set that has the given type.

filterTermsByReferentIdHavingTypeImpl :: TypeReference -> Set Id -> Transaction (Set Id)

Return the subset of the given set that has the given type.

termReferentsByPrefix :: ShortHash -> Transaction (Set Id)

Get the set of user-defined terms-or-constructors whose hash matches the given prefix.

withConnection :: forall x. (Connection -> m x) -> m x

Acquire a new connection to the same underlying database file this codebase object connects to.

withConnectionIO :: forall x. (Connection -> IO x) -> IO x

Acquire a new connection to the same underlying database file this codebase object connects to.

preloadBranch :: CausalHash -> m ()

This optimization allows us to pre-fetch a branch from SQLite into the branch cache when we know we'll need it soon, but not immediately. E.g. the user has switched a branch, but hasn't run any commands on it yet.

This combinator returns immediately, but warms the cache in the background with the desired branch.

type CodebasePath = FilePath

A directory that contains a codebase.

data LocalOrRemote

Whether a codebase is local or remote.

Constructors

Local

Remote