Unison.Hashing.V2.Convert

Documentation

type ResolutionResult v a r = Either (Seq (ResolutionFailure v a)) r

hashBranch0 :: Branch0 m -> Hash

hashCausal :: Hashable e => e -> Set (RawHash h) -> Hash

hashDataDecls :: Var v => Map v (DataDeclaration v a) -> ResolutionResult v a [(v, Id, DataDeclaration v a)]

hashDecls :: Var v => Map v (Decl v a) -> ResolutionResult v a [(v, Id, Decl v a)]

hashPatch :: Patch -> Hash

hashClosedTerm :: Var v => Term v a -> Id

hashTermComponents :: forall v a. Var v => Map v (Term v a, Type v a) -> Map v (Id, Term v a, Type v a)

hashTermComponentsWithoutTypes :: forall v a. Var v => Map v (Term v a) -> Map v (Id, Term v a)

This shouldn't be used when storing terms in the codebase, as it doesn't incorporate the type into the hash. this should only be used in cases where you just need a way to identify some terms that you have, but won't be saving them.

typeToReference :: Var v => Type v a -> Reference

typeToReferenceMentions :: Var v => Type v a -> Set Reference