Unison.Hashing.V2.Convert

Description

Documentation

type ResolutionResult a r = Either (Seq (ResolutionFailure a)) r

hashBranch0 :: Branch0 m -> Hash

hashCausal :: ContentAddressable e => e -> Set CausalHash -> (CausalHash, HashFor e)

hashDataDecls :: Var v => Map v (DataDeclaration v a) -> ResolutionResult a [(v, Id, DataDeclaration v a)]

hashDecls :: Var v => Map v (Decl v a) -> ResolutionResult a [(v, Id, Decl v a)]

hashPatch :: Patch -> Hash

hashClosedTerm :: Var v => Term v a -> Id

hashTermComponents :: forall v a extra. Var v => Map v (Term v a, Type v a, extra) -> Map v (TermReferenceId, Term v a, Type v a, extra)

hashTermComponentsWithoutTypes :: forall v a. Var v => Map v (Term v a) -> Map v (TermReferenceId, Term v a)

This shouldn't be used when storing terms in the codebase, as it doesn't incorporate the type into the hash. this should only be used in cases where you just need a way to identify some terms that you have, but won't be saving them.

typeToReference :: Var v => Type v a -> Reference

typeToReferenceMentions :: Var v => Type v a -> Set Reference