Unison.Hashing.V2.Hashable

Documentation

class Hashable t where

This typeclass provides a mechanism for obtaining a content-based hash for Unison types & terms. Be wary that Unison requires that these hashes be deterministic, any change to a Hashable instance requires a full codebase migration and should not be taken lightly.

Methods

hash :: t -> Hash