Unison.KindInference.Solve.Monad

Documentation

newtype Solve v loc a

Constructors

Solve

Fields

unSolve :: Env -> SolveState v loc -> (a, SolveState v loc)

data Env

Constructors

Env

Fields

prettyPrintEnv :: PrettyPrintEnv

data SolveState v loc

The SolveState holds all kind constraints gathered for each type. For example, after processing data and effect decls the typeMap will hold entries for every decl, and looking up the corresponding UVar in constraints will return its kind.

The other fields, unifVars and newUnifVars, are relevant when interleaving constraint generation with solving. Constraint generation needs to create fresh unification variables, so it needs the set of bound unification variables from unifVars. newUnifVars holds the uvars that are candidates for kind defaulting (see defaultUnconstrainedVars).

Constructors

SolveState

Fields

unifVars :: !(Set Symbol)

newUnifVars :: [UVar v loc]

constraints :: !(UFMap (UVar v loc) (Descriptor v loc))

typeMap :: !(Map (Type v loc) (NonEmpty (UVar v loc)))

data Descriptor v loc

Constraints associated with a unification variable

Constructors

Descriptor

Fields

descriptorConstraint :: Maybe (Constraint (UVar v loc) v loc)

type ConstraintMap v loc = UFMap (UVar v loc) (Descriptor v loc)

run :: Env -> SolveState v loc -> Solve v loc a -> (a, SolveState v loc)

Runner for the Solve monad

emptyState :: SolveState v loc

Initial solve state

find :: Var v => UVar v loc -> Solve v loc (Maybe (Constraint (UVar v loc) v loc))

Lookup the constraints associated with a unification variable

genStateL :: Lens' (SolveState v loc) (GenState v loc)

Helper for inteleaving constraint generation and solving

runGen :: Var v => Gen v loc a -> Solve v loc a

Interleave constraint generation into constraint solving

addUnconstrainedVar :: Var v => UVar v loc -> Solve v loc ()

Add a unification variable to the constarint mapping with no constraints. This is done on uvars created during constraint generation to initialize the new uvars (see runGen).