Unison.PrettyPrintEnv.MonadPretty

Documentation

type MonadPretty v m = (Var v, MonadReader (Env v) m)

data Env v

Constructors

Env

Fields

boundTerms :: !(Set v)

boundTypes :: !(Set v)

freeTerms :: !(Set v)

ppe :: !PrettyPrintEnv

runPretty :: Var v => PrettyPrintEnv -> Reader (Env v) a -> a

addTypeVars :: MonadPretty v m => [v] -> m a -> m a

Add type variables to the set of variables that need to be avoided

willCaptureType :: MonadPretty v m => [v] -> m Bool

Check if a list of type variables contains any variables that need to be avoided

withBoundTerm :: MonadPretty v m => v -> m a -> m a

withBoundTerms :: MonadPretty v m => [v] -> m a -> m a