Unison.PrettyPrintEnv.MonadPretty

Documentation

type MonadPretty v m = (Var v, MonadReader (PrettyPrintEnv, Set v) m)

getPPE :: MonadPretty v m => m PrettyPrintEnv

withPPE :: MonadPretty v m => PrettyPrintEnv -> m a -> m a

Run a computation with a modified PrettyPrintEnv, restoring the original

applyPPE :: MonadPretty v m => (PrettyPrintEnv -> a) -> m a

applyPPE2 :: MonadPretty v m => (PrettyPrintEnv -> a -> b) -> a -> m b

applyPPE3 :: MonadPretty v m => (PrettyPrintEnv -> a -> b -> c) -> a -> b -> m c

modifyPPE :: MonadPretty v m => (PrettyPrintEnv -> PrettyPrintEnv) -> m a -> m a

Run a computation with a modified PrettyPrintEnv, restoring the original

modifyTypeVars :: MonadPretty v m => (Set v -> Set v) -> m a -> m a

addTypeVars :: MonadPretty v m => [v] -> m a -> m a

Add type variables to the set of variables that need to be avoided

willCapture :: MonadPretty v m => [v] -> m Bool

Check if a list of type variables contains any variables that need to be avoided

runPretty :: Var v => PrettyPrintEnv -> Reader (PrettyPrintEnv, Set v) a -> a