Unison.Typechecker

Description

This module is the primary interface to the Unison typechecker module Unison.Typechecker (admissibleTypeAt, check, check', checkAdmissible', equals, locals, subtype, isSubtype, synthesize, synthesize', typeAt, wellTyped) where

Documentation

type Name = Text

data Notes v loc

Constructors

Notes

Fields

bugs :: Seq (CompilerBug v loc)

errors :: Seq (ErrorNote v loc)

infos :: Seq (InfoNote v loc)

convertResult :: Result v loc a -> Result (Notes v loc) a

data NamedReference v loc

Constructors

NamedReference

Fields

fqn :: Name

fqnType :: Type v loc

replacement :: Either v Referent

data Env v loc

Constructors

Env

Fields

_ambientAbilities :: [Type v loc]

_typeLookup :: TypeLookup v loc

_termsByShortname :: Map Name [NamedReference v loc]

typeLookup :: forall v loc. Lens' (Env v loc) (TypeLookup v loc)

termsByShortname :: forall v loc. Lens' (Env v loc) (Map Name [NamedReference v loc])

ambientAbilities :: forall v loc. Lens' (Env v loc) [Type v loc]

synthesize :: (Monad f, Var v, Ord loc) => Env v loc -> Term v loc -> ResultT (Notes v loc) f (Type v loc)

Infer the type of a Term, using a function to resolve the type of Ref constructors contained in that term.

isSubtype :: Var v => Type v loc -> Type v loc -> Bool

isEqual :: Var v => Type v loc -> Type v loc -> Bool

type TDNR f v loc a = StateT (Term v loc) (ResultT (Notes v loc) f) a

data Resolution v loc

Constructors

Resolution

Fields

resolvedName :: Text

inferredType :: Type v loc

resolvedLoc :: loc

suggestions :: [Suggestion v loc]

synthesizeAndResolve :: (Monad f, Var v, Monoid loc, Ord loc) => Env v loc -> TDNR f v loc (Type v loc)

Infer the type of a Term, using type-directed name resolution to attempt to resolve unknown symbols.

compilerBug :: CompilerBug v loc -> Result (Notes v loc) ()

typeError :: ErrorNote v loc -> Result (Notes v loc) ()

btw :: Monad f => InfoNote v loc -> ResultT (Notes v loc) f ()

liftResult :: Monad f => Result (Notes v loc) a -> TDNR f v loc a

typeDirectedNameResolution :: forall v loc f. (Monad f, Var v, Ord loc, Monoid loc) => Notes v loc -> Type v loc -> Env v loc -> TDNR f v loc (Type v loc)

check :: (Monad f, Var v, Ord loc) => Env v loc -> Term v loc -> Type v loc -> ResultT (Notes v loc) f (Type v loc)

Check whether a term matches a type, using a function to resolve the type of Ref constructors contained in the term. Returns typ if successful, and a note about typechecking failure otherwise.

wellTyped :: (Monad f, Var v, Ord loc) => Env v loc -> Term v loc -> f Bool

checkAdmissible e t` tests that `(f : t -> r) e` is well-typed. If t has quantifiers, these are moved outside, so if `t : forall a . a`, this will check that `(f : forall a . a -> a) e` is well typed. checkAdmissible' :: Var v => Term v -> Type v -> Either Note (Type v) checkAdmissible' term typ = synthesize' (Term.blank() ann_ tweak typ app_ term) where tweak (Type.ForallNamed' v body) = Type.forall() v (tweak body) tweak t = Type.arrow() t t | Returns True if the expression is well-typed, False otherwise